% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{as.list.myClimList}
\alias{as.list.myClimList}
\title{Function convert myClimList to list §§§}
\usage{
\method{as.list}{myClimList}(x)
}
\arguments{
\item{x}{myClim object (see \link{myClim-package})}
}
\description{
Function convert myClimList to list §§§
}
