% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data.source_data_formats}
\alias{data.source_data_formats}
\title{Formats of source data files}
\format{
An object of class \code{list} of length 3.
}
\usage{
data.source_data_formats
}
\description{
Formats of source data files
}
\keyword{datasets}
