% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{DataFormat-class}
\alias{DataFormat-class}
\title{Class for source file data format}
\description{
Class for source file data format
}
\section{Slots}{

\describe{
\item{\code{has_header}}{columns separator}

\item{\code{separator}}{columns separator}

\item{\code{date_column}}{index of date column}

\item{\code{date_format}}{format of date}
}}

