% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{length.myClimList}
\alias{length.myClimList}
\title{Length function for myClim object}
\usage{
\method{length}{myClimList}(x, ...)
}
\arguments{
\item{x}{myClim object see \link{myClim-package}}

\item{...}{other parameters from function length}
}
\description{
Function return number of localities.
}
\examples{
length(mc_data_example_agg)
}
