% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{mc_agg}
\alias{mc_agg}
\title{Agregate data by function}
\usage{
mc_agg(
  data,
  fun = NULL,
  period = NULL,
  use_utc = TRUE,
  percentiles = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{data}{in format for preparing or calculation}

\item{fun}{aggregation function ("min", "max", "mean", "percentile", "sum", "count", "coverage")
Can be character vector of function names or list. if NULL than no aggregation.
\itemize{
\item functions are applied to all sensors. Sensors aren't renamed.
\item Names of items in list are sensor_names and items are vectors of functions applied to sensors.
Names of new sensors are in format {sensor_name}_{function}.
}

function coverage is count_values/count_all_records}

\item{period}{of aggregation - same as breaks in cut.POSIXt; if NULL then no aggregation

start day of week is monday}

\item{use_utc}{if set FALSE then datetime changed by locality tz_offset (default TRUE);
Non-UTC time can by used only for period \code{day} and bigger.}

\item{percentiles}{\itemize{
\item vector of percentile numbers; numbers are from range 0-100; every number generate new sensor
}}

\item{na.rm}{parameter for aggregation function; It isn't used for count and coverage.}
}
\value{
aggregated data in format for calculating
}
\description{
Function create aggregated data in format for calculation. If fun is NULL and period is NULL, than
function only convert source data to format for calculation.
}
\details{
If first or last period isn't full filled, data are cropped and a warning is shown. New sensors have
same sensor_id as source one. It is usefull for detecting source sensor. Sensors without data are excluded.
Aggregation functions return NA for empty vector. Except count, it return 0.
}
\examples{
example_cleaned_tomst_data <- mc_agg(example_cleaned_tomst_data, c(min, max, percentile), "hour", percentiles = 50, na.rm=TRUE)
}
