% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mc_calc_agg}
\alias{mc_calc_agg}
\title{Agregate data by function}
\usage{
mc_calc_agg(
  data,
  fun,
  breaks,
  localities = NULL,
  sensors = NULL,
  use_utc = F,
  suffix = "_agg",
  ...
)
}
\arguments{
\item{data}{in format for calculation}

\item{fun}{aggregation function}

\item{breaks}{cut function parameter}

\item{localities}{locality_ids for filtering data; if empty then all}

\item{sensors}{sensor_ids for filtering data; if empty then all}

\item{use_utc}{if set FALSE then datetime changed by locality tz_offset (default FALSE)}

\item{suffix}{of new locality name}

\item{...}{parameters for aggregation function}
}
\value{
aggregated data in standard format
}
\description{
Function add agregated locality.
}
\examples{
example_cleaned_tomst_data <- mc_calc_agg(example_cleaned_tomst_data, quantile, "hour", probs = 0.5, na.rm=TRUE)
}
