% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mc_calc_agg_mean}
\alias{mc_calc_agg_mean}
\title{Agregate data by mean function}
\usage{
mc_calc_agg_mean(
  data,
  breaks,
  localities = NULL,
  sensors = NULL,
  use_utc = F,
  ...
)
}
\arguments{
\item{data}{all data in standard format}

\item{breaks}{cut function parameter}

\item{localities}{locality_ids for filtering data; if empty then all}

\item{sensors}{sensor_ids for filtering data; if empty then all}

\item{use_utc}{if set FALSE then datetime changed by locality tz_offset; default FALSE}

\item{...}{parameters for mean function}
}
\value{
aggregated data in standard format
}
\description{
Function return aggregated data by mean
}
\examples{
example_cleaned_tomst_data <- mc_calc_agg_mean(example_cleaned_tomst_data, "hour", na.rm=TRUE)
}
