% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mc_calc_cumsum}
\alias{mc_calc_cumsum}
\title{Cumulative sum}
\usage{
mc_calc_cumsum(data, sensors, output_suffix = "_cumsum", localities = NULL)
}
\arguments{
\item{data}{myClim object in Calc-format see \code{\link[=mc_agg]{mc_agg()}} and \link{myClim-package}}

\item{sensors}{names of sensors for which is calculated cumulative sum}

\item{output_suffix}{name suffix for new names (default "_cumsum")}

\item{localities}{list of locality_ids for calculation; if NULL then all (default NULL)}
}
\value{
The same myClim object as input but with added cumsum sensors.||
}
\description{
||Function calculate cumulative sum for sensors and add output as new sensors.
Names of new sensors are original sensor name + \code{outpus_suffix}.
}
\details{
If value type of sensor is logical, then output type is integer.
}
\examples{
cumsum_data <- mc_calc_cumsum(mc_data_example_calc, c("TMS_T1", "TMS_T2"))
}
