% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mc_calc_fdd}
\alias{mc_calc_fdd}
\title{Freezing Degree Days}
\usage{
mc_calc_fdd(data, sensor, output_prefix = "FDD", t_base = 0, localities = NULL)
}
\arguments{
\item{data}{in format for calculation}

\item{sensor}{name of temperature sensor}

\item{output_prefix}{name prefix of new FDD sensor (default "FDD")

name of output sensor consists of output_prefix and value t_base}

\item{t_base}{threshold temperaturefor calculation FDD (default 0)}

\item{localities}{list of locality_ids for calculation; if NULL then all (default NULL)}
}
\value{
input data with added FDD sensor
}
\description{
Function add new virtual sensor with values of FDD Freezing Degree Days.
}
\details{
Maximal step length of data is day.
}
