% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mc_calc_snow}
\alias{mc_calc_snow}
\title{Snow detection}
\usage{
mc_calc_snow(
  data,
  sensor,
  output_sensor = "snow",
  localities = NULL,
  dr = 2,
  tmax = 0.5
)
}
\arguments{
\item{data}{in format for calculation}

\item{sensor}{name of temperature sensor}

\item{output_sensor}{name of new snow sensor (default "snow")}

\item{localities}{names for calculation; if empty then all}

\item{dr}{delta range}

\item{tmax}{maximal temperature}
}
\value{
input data with added snow sensor
}
\description{
Function add sensor to locality with snow detection
}
\examples{
snow <- mc_calc_snow(example_tomst_data1, "TMS_T3")
}
