% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mc_calc_tomst_dendro}
\alias{mc_calc_tomst_dendro}
\title{Converting raw values of TOMST dendrometer to micrometers}
\usage{
mc_calc_tomst_dendro(
  data,
  dendro_sensor = myClim:::.model_const_SENSOR_DEND_TOMST_r_delta,
  output_sensor = myClim:::.model_const_SENSOR_r_delta,
  localities = NULL
)
}
\arguments{
\item{data}{myClim object in Calc-format see \code{\link[=mc_agg]{mc_agg()}} and \link{myClim-package}}

\item{dendro_sensor}{name of radius difference sensor to be converted from raw to micrometers (default "DEND_TOMST_r_delta") see \code{names(mc_data_sensors)}}

\item{output_sensor}{name of new snow sensor (default "r_delta")}

\item{localities}{list of locality_ids for calculation; if NULL then all (default NULL)}
}
\value{
myClim object same as input but with added r_delta sensor
}
\description{
This function convert differences of trunk radius fram raw TOMST units to micrometers.
}
\examples{
calc_data <- mc_calc_tomst_dendro(mc_data_example_calc, localities="A1E05")
}
