% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mc_calc_vpd}
\alias{mc_calc_vpd}
\title{Calculate vapor pressure deficit (in kPa)}
\usage{
mc_calc_vpd(
  data,
  temp_sensor = "HOBO_T_C",
  rh_sensor = "HOBO_RH",
  output_sensor = "VPD",
  altitude = 0,
  metadata_altitude = TRUE,
  localities = NULL
)
}
\arguments{
\item{data}{myClim object in cleaned Prep-format or Calc-formt see \code{\link[=mc_agg]{mc_agg()}} and \link{myClim-package}}

\item{temp_sensor}{name of temperature sensor

Temperature sensor must be in T_C physical.}

\item{rh_sensor}{name relative humidity sensor

Temperature sensor must be in RH_perc physical.}

\item{output_sensor}{name of new snow sensor (default "VPD")}

\item{altitude}{value in meters (default 0)}

\item{metadata_altitude}{if TRUE then altitude from metadata of locality is used (default TRUE)}

\item{localities}{list of locality_ids for calculation; if NULL then all (default NULL)}
}
\value{
myClim object same as input but with added VPD sensor
}
\description{
Function calculate vapor pressure deficit (in kPa) from temperature and relative humidity readings using equation
from the CR-5 Users Manual 2009–12 from Buck Research modified from Buck (1981) and adapted by Jones, 2013 (eq. 5.15)
Elevation to pressure conversion function uses eq. 3.7 from Campbell G.S. & Norman J.M. (1998).
}
\examples{
calc_data <- mc_calc_vpd(mc_data_example_calc, "HOBO_T_C", "HOBO_RH", localities="A2E32")
}
\references{
Jones H.G. (2014) Plants and Microclimate, Third Edit. Cambridge University Press, Cambridge Buck A.L. (1981)
New equations for computing vapor pressure and enhancment factor. Journal of Applied Meteorology 20: 1527–1532.
Campbell G.S. & Norman J.M. (1998). An Introduction to Environmental Biophysics, Springer New York, New York, NY
}
