% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{mc_clean_crop}
\alias{mc_clean_crop}
\title{Crop datetime}
\usage{
mc_clean_crop(data, start = NULL, end = NULL)
}
\arguments{
\item{data}{in standard format}

\item{start}{POSIXct datetime in UTC; is optional}

\item{end}{POSIXct datetime in UTC; is optional}
}
\value{
cropped data in standard format
}
\description{
This function crop data by datetime
}
\examples{
cleaned_example_tomst_data1 <- mc_clean_crop(example_tomst_data1, end=as.POSIXct("2020-02-01"))
}
