% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{mc_clean_datetime_step}
\alias{mc_clean_datetime_step}
\title{Cleaning datetime series}
\usage{
mc_clean_datetime_step(data)
}
\arguments{
\item{data}{character data in standard format}
}
\value{
cleaned data in standard format
}
\description{
This function change datetime and values series. Result series has constant
step without duplicits and missed values are filled in as NA.
}
\examples{
cleaned_example_tms_data1 <- mc_clean_datetime_step(example_tms_data1)
}
