% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{mc_clean_logs}
\alias{mc_clean_logs}
\title{Get all clean log messages}
\usage{
mc_clean_logs(data)
}
\arguments{
\item{data}{character data in standard format}
}
\value{
dataframe with columns locality_id, serial_number, clean_type, message
}
\description{
This function return dataframe with all clean log messages
}
\examples{
cleaned_example_tomst_data1 <- mc_clean_datetime_step(example_tomst_data1)
}
