% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{mc_clean_solar_tz}
\alias{mc_clean_solar_tz}
\title{Solar TZ offset}
\usage{
mc_clean_solar_tz(data)
}
\arguments{
\item{data}{character data in standard format}
}
\value{
data with changed TZ offset in standard format
}
\description{
This function compute TZ offset in localities by solar time
}
\examples{
cleaned_example_tomst_data1 <- mc_clean_solar_tz(cleaned_example_tomst_data1)
}
