% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{mc_clean_user_tz}
\alias{mc_clean_user_tz}
\title{Set user defined TZ offset}
\usage{
mc_clean_user_tz(data, tz_offsets)
}
\arguments{
\item{data}{in standard format}

\item{tz_offsets}{named list (name: locality_id, item: tz_offset in rounded minutes)}
}
\value{
data with changed TZ offset in standard format
}
\description{
This function set user defined TZ offsets in localities
}
\examples{
example_tomst_data2 <- mc_clean_solar_tz(example_tomst_data2, list(None=60))
}
