% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\docType{data}
\name{mc_const_CALIB_MOIST_WCOR_T}
\alias{mc_const_CALIB_MOIST_WCOR_T}
\title{default temperature drift correction parameter for water for - TMS moisture sensor}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
mc_const_CALIB_MOIST_WCOR_T
}
\description{
default temperature drift correction parameter for water for - TMS moisture sensor
}
\keyword{datasets}
