% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_coverage}
\alias{mc_const_SENSOR_coverage}
\title{Coverage sensor id see \code{\link[=mc_agg]{mc_agg()}}}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_coverage
}
\description{
Coverage sensor id see \code{\link[=mc_agg]{mc_agg()}}
}
\keyword{datasets}
