% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_dendro_l_um}
\alias{mc_const_SENSOR_dendro_l_um}
\title{Radius difference sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_dendro_l_um
}
\description{
Radius difference sensor id
}
\keyword{datasets}
