% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_Dendro_raw}
\alias{mc_const_SENSOR_Dendro_raw}
\title{TOMST Dendrometer radius difference sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_Dendro_raw
}
\description{
TOMST Dendrometer radius difference sensor id
}
\keyword{datasets}
