% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_Dendro_T}
\alias{mc_const_SENSOR_Dendro_T}
\title{TOMST Dendrometer temperature sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_Dendro_T
}
\description{
TOMST Dendrometer temperature sensor id
}
\keyword{datasets}
