% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_FDD}
\alias{mc_const_SENSOR_FDD}
\title{Freezing Degree Days sensor id see \code{\link[=mc_calc_fdd]{mc_calc_fdd()}}}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_FDD
}
\description{
Freezing Degree Days sensor id see \code{\link[=mc_calc_fdd]{mc_calc_fdd()}}
}
\keyword{datasets}
