% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_GDD}
\alias{mc_const_SENSOR_GDD}
\title{Growing Degree Days sensor id see \code{\link[=mc_calc_gdd]{mc_calc_gdd()}}}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_GDD
}
\description{
Growing Degree Days sensor id see \code{\link[=mc_calc_gdd]{mc_calc_gdd()}}
}
\keyword{datasets}
