% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_HOBO_EXTT}
\alias{mc_const_SENSOR_HOBO_EXTT}
\title{Onset HOBO external temperature sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_HOBO_EXTT
}
\description{
Onset HOBO external temperature sensor id
}
\keyword{datasets}
