% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_integer}
\alias{mc_const_SENSOR_integer}
\title{General integer sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_integer
}
\description{
General integer sensor id
}
\keyword{datasets}
