% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_logical}
\alias{mc_const_SENSOR_logical}
\title{General logical sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_logical
}
\description{
General logical sensor id
}
\keyword{datasets}
