% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_MINIKIN_T}
\alias{mc_const_SENSOR_MINIKIN_T}
\title{EMS Minikin temperature sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_MINIKIN_T
}
\description{
EMS Minikin temperature sensor id
}
\keyword{datasets}
