% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_precipitation}
\alias{mc_const_SENSOR_precipitation}
\title{Precipitation sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_precipitation
}
\description{
Precipitation sensor id
}
\keyword{datasets}
