% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_QTi_PPFD}
\alias{mc_const_SENSOR_QTi_PPFD}
\title{EMS Minikin QTi PPFD sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_QTi_PPFD
}
\description{
EMS Minikin QTi PPFD sensor id
}
\keyword{datasets}
