% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_real}
\alias{mc_const_SENSOR_real}
\title{General real sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_real
}
\description{
General real sensor id
}
\keyword{datasets}
