% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_snow_bool}
\alias{mc_const_SENSOR_snow_bool}
\title{Snow existence sensor id see \code{\link[=mc_calc_snow]{mc_calc_snow()}}}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_snow_bool
}
\description{
Snow existence sensor id see \code{\link[=mc_calc_snow]{mc_calc_snow()}}
}
\keyword{datasets}
