% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_snow_fresh}
\alias{mc_const_SENSOR_snow_fresh}
\title{Height of newly fallen snow sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_snow_fresh
}
\description{
Height of newly fallen snow sensor id
}
\keyword{datasets}
