% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_snow_total}
\alias{mc_const_SENSOR_snow_total}
\title{Height snow sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_snow_total
}
\description{
Height snow sensor id
}
\keyword{datasets}
