% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_sun_shine}
\alias{mc_const_SENSOR_sun_shine}
\title{Time of sun shine sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_sun_shine
}
\description{
Time of sun shine sensor id
}
\keyword{datasets}
