% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_T_C}
\alias{mc_const_SENSOR_T_C}
\title{Temperature sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_T_C
}
\description{
Temperature sensor id
}
\keyword{datasets}
