% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_TH2_RH}
\alias{mc_const_SENSOR_TH2_RH}
\title{EMS Minikin TH2 RH sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_TH2_RH
}
\description{
EMS Minikin TH2 RH sensor id
}
\keyword{datasets}
