% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_TH2_T}
\alias{mc_const_SENSOR_TH2_T}
\title{EMS Minikin TH2 temperature sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_TH2_T
}
\description{
EMS Minikin TH2 temperature sensor id
}
\keyword{datasets}
