% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_Thermo_T}
\alias{mc_const_SENSOR_Thermo_T}
\title{Default sensor for TOMST Thermologger temperature}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_Thermo_T
}
\description{
Default sensor for TOMST Thermologger temperature
}
\keyword{datasets}
