% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_TMS_moist}
\alias{mc_const_SENSOR_TMS_moist}
\title{Default sensor for TOMST TMS raw soil moisture}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_TMS_moist
}
\description{
This constant is used in the function \link{mc_calc_vwc} as default for sensor for
converting the raw TMS soil moisture (scaled TDT signal) to volumetric water content (VWC).
mc_const_SENSOR_TMS_moist = "TMS_moist"
}
\keyword{datasets}
