% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_TMS_T2}
\alias{mc_const_SENSOR_TMS_T2}
\title{TMS temperature sensor id - surface of soil}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_TMS_T2
}
\description{
TMS temperature sensor id - surface of soil
}
\keyword{datasets}
