% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_TMS_T3}
\alias{mc_const_SENSOR_TMS_T3}
\title{TMS temperature sensor id - air}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_TMS_T3
}
\description{
TMS temperature sensor id - air
}
\keyword{datasets}
