% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_VPD}
\alias{mc_const_SENSOR_VPD}
\title{Vapor Pressure Deficit sensor id see \code{\link[=mc_calc_vpd]{mc_calc_vpd()}}}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_VPD
}
\description{
Vapor Pressure Deficit sensor id see \code{\link[=mc_calc_vpd]{mc_calc_vpd()}}
}
\keyword{datasets}
