% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_VWC}
\alias{mc_const_SENSOR_VWC}
\title{Volumetric soil moisture sensor id see \code{\link[=mc_calc_vwc]{mc_calc_vwc()}}}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_VWC
}
\description{
Volumetric soil moisture sensor id see \code{\link[=mc_calc_vwc]{mc_calc_vwc()}}
}
\keyword{datasets}
