% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_wind_speed}
\alias{mc_const_SENSOR_wind_speed}
\title{Speed of wind sensor id}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_wind_speed
}
\description{
Speed of wind sensor id
}
\keyword{datasets}
