% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_data_example_calc}
\alias{mc_data_example_calc}
\title{Example data in calc-format and 15 minutes step.}
\format{
An object of class \code{myClimList} (inherits from \code{list}) of length 2.
}
\usage{
mc_data_example_calc
}
\description{
Example data in calc-format and 15 minutes step.
}
\keyword{datasets}
