% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_data_example_clean}
\alias{mc_data_example_clean}
\title{Example cleaned data in prep-format.}
\format{
An object of class \code{myClimList} (inherits from \code{list}) of length 2.
}
\usage{
mc_data_example_clean
}
\description{
Example cleaned data in prep-format.
}
\keyword{datasets}
