% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_data_example_raw}
\alias{mc_data_example_raw}
\title{Example data in Raw-format. Data aren't cleaned.}
\format{
An object of class \code{myClimList} (inherits from \code{list}) of length 2.
}
\usage{
mc_data_example_raw
}
\description{
Example data in Raw-format. Data aren't cleaned.
}
\keyword{datasets}
