% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_data_formats}
\alias{mc_data_formats}
\title{Formats of source data files}
\format{
An object of class \code{environment} of length 2.
}
\usage{
mc_data_formats
}
\description{
Formats of source data files
}
\keyword{datasets}
