% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_data_vwc_parameters}
\alias{mc_data_vwc_parameters}
\title{Volumetric water content parameters}
\format{
An object of class \code{data.frame} with 13 rows and 9 columns.
}
\usage{
mc_data_vwc_parameters
}
\description{
data.frame with columns:
\itemize{
\item soiltype
\item a
\item b
\item c
\item rho
\item clay
\item silt
\item sand
\item ref
}
}
\keyword{datasets}
