% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_DataFormat-class}
\alias{mc_DataFormat-class}
\alias{mc_DataFormat}
\title{Class for source file data format}
\description{
By this class is parsed source data file.
}
\section{Slots}{

\describe{
\item{\code{has_header}}{detects if table contains header (default TRUE)}

\item{\code{separator}}{columns separator (default ";")}

\item{\code{date_column}}{index of date column (default NA)}

\item{\code{date_format}}{format of date (default NA)}

\item{\code{na_strings}}{strings for NA values (default NA)}

\item{\code{columns}}{list with names and indexes of value columns (default list())}

\item{\code{filename_serial_number_pattern}}{character pattern for detecting serial_number from filename (default NA)}

\item{\code{data_row_pattern}}{character pattern for detecting right file format}

\item{\code{logger_type}}{type of logger: TMS, ThermoDatalogger (default NA)}
}}

