% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.R
\name{mc_eco_agg}
\alias{mc_eco_agg}
\title{Agregate data by function}
\usage{
mc_eco_agg(data, fun, breaks, localities = c(), sensors = c(), ...)
}
\arguments{
\item{data}{all data in standard format}

\item{fun}{aggregation function}

\item{localities}{locality_ids for filtering data; if empty then all}

\item{braks}{cut function parameter}

\item{snesors}{sensor_ids for filtering data; if empty then all}
}
\value{
aggregated data in standard formta
}
\description{
Function return aggregated data by function
}
\examples{
example_cleaned_tomst_data <- mc_eco_agg(example_cleaned_tomst_data, quantile, "hour", probs = 0.5, na.rm=TRUE)
}
