% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.R
\name{mc_eco_agg_quantile}
\alias{mc_eco_agg_quantile}
\title{Agregate data by quantile function}
\usage{
mc_eco_agg_quantile(
  data,
  breaks,
  probs,
  localities = NULL,
  sensors = NULL,
  ...
)
}
\arguments{
\item{data}{all data in standard format}

\item{breaks}{cut function parameter}

\item{probs}{value 0-1}

\item{localities}{locality_ids for filtering data; if empty then all}

\item{sensors}{sensor_ids for filtering data; if empty then all}
}
\value{
aggregated data in standard format
}
\description{
Function return aggregated data by quantile function
}
\examples{
example_cleaned_tomst_data <- mc_eco_agg_quantile(example_cleaned_tomst_data, "hour", 0.1, na.rm=TRUE)
}
