% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.R
\name{mc_eco_snow}
\alias{mc_eco_snow}
\title{Snow detection}
\usage{
mc_eco_snow(
  data,
  sensor,
  localities = c(),
  dr = 2,
  tmax = 0.5,
  interval_length = 15
)
}
\arguments{
\item{data}{all data in standard format}

\item{sensor}{name of temperature sensor}

\item{localities}{names of localities; if empty then all}

\item{dr}{delta range}

\item{tmax}{maximal temperature}

\item{interval_length}{length of interval in minutes (default 15)}
}
\value{
data.frame with datetime column and logical columns named by serial_number of loggers
}
\description{
Function detect snow based on detrended time series
}
\examples{
mc_eco_snow(example_tms_data1, "T3")
}
