% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.R
\name{mc_eco_snow_agg}
\alias{mc_eco_snow_agg}
\title{Snow detection summary}
\usage{
mc_eco_snow_agg(snow_data, period = 3)
}
\arguments{
\item{snow_data}{data from function mc_eco_snow}

\item{period}{count days for continuous cover of snow (default 3)}
}
\value{
data.frame with columns serial_number, snow_days, first_day, last_day, first_day_period, last_day_period
}
\description{
Function return summary info about snow detection
}
\examples{
data <- mc_eco_snow(example_tms_data1, "T3")
mc_eco_snow_agg(data)
}
