% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{mc_env_moist}
\alias{mc_env_moist}
\title{Moisture environment variables}
\usage{
mc_env_moist(
  data,
  period,
  use_utc = TRUE,
  custom_start = NULL,
  custom_end = NULL
)
}
\arguments{
\item{data}{myClim object in cleaned Raw-format or Calc-formt see \code{\link[=mc_agg]{mc_agg()}} and \link{myClim-package}}

\item{period}{output period see \code{\link[=mc_agg]{mc_agg()}}}

\item{use_utc}{if FALSE, then local time is used for day aggregation see \code{\link[=mc_agg]{mc_agg()}}}

\item{custom_start}{start date for custom period see \code{\link[=mc_agg]{mc_agg()}}}

\item{custom_end}{end date for custom period see \code{\link[=mc_agg]{mc_agg()}}}
}
\value{
table in long format with environment variables
}
\description{
Function ...
}
\details{
\itemize{
\item used all VWC sensors
\item functions
\itemize{
\item 5p: fifth percentile
\item mean: mean
\item 95p: ninety-fifth percentile
\item sd: standard deviation
}
}
}
\examples{
data <- mc_prep_crop(mc_data_example_calc, lubridate::ymd_h("2020-11-01 00"), lubridate::ymd_h("2021-02-01 00"), end_included = FALSE)
data <- mc_calc_vwc(data, localities=c("A2E32", "A6W79"))
mc_env_moist(data, "month")
}
