% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{mc_env_temp}
\alias{mc_env_temp}
\title{Tempereature environment variables}
\usage{
mc_env_temp(
  data,
  period = NULL,
  use_utc = TRUE,
  custom_start = NULL,
  custom_end = NULL,
  gdd_t_base = 5,
  fdd_t_base = 0
)
}
\arguments{
\item{data}{myClim object in cleaned Prep-format or Calc-formt see \code{\link[=mc_agg]{mc_agg()}} and \link{myClim-package}}
}
\value{
long table with environment variables
}
\description{
Function ...
}
