% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{mc_env_vpd}
\alias{mc_env_vpd}
\title{vapor pressure deficit environment variables}
\usage{
mc_env_vpd(
  data,
  period,
  use_utc = TRUE,
  custom_start = NULL,
  custom_end = NULL
)
}
\arguments{
\item{data}{cleaned myClim object see \link{myClim-package}}

\item{period}{output period see \code{\link[=mc_agg]{mc_agg()}}}

\item{use_utc}{if FALSE, then local time is used for day aggregation see \code{\link[=mc_agg]{mc_agg()}}}

\item{custom_start}{start date for custom period see \code{\link[=mc_agg]{mc_agg()}}}

\item{custom_end}{end date for custom period see \code{\link[=mc_agg]{mc_agg()}}}
}
\value{
table in long format with environment variables
}
\description{
Function ...
}
\details{
\itemize{
\item used all VPD sensors
\item functions
\itemize{
\item mean: mean of daily mean
\item max95p: ninety-fifth percentile of daily maximum
}
}
}
