% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{mc_feed_files}
\alias{mc_feed_files}
\title{Reading files}
\usage{
mc_feed_files(files, logger_type)
}
\arguments{
\item{files}{vector of character - files with data}

\item{logger_type}{character - type of logger (TMS)}
}
\value{
data in standard format
}
\description{
This function read data files of one logger type. Locality is set None.
}
\examples{
example_tms_data <- microclim::mc_feed_files(c("examples/data/TMS/data_91184101_0.csv", "examples/data/TMS/data_94184102_0.csv"), "TMS")
}
