% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{mc_feed_from_csv}
\alias{mc_feed_from_csv}
\title{Data files reading by CSV}
\usage{
mc_feed_from_csv(csv_files_table, csv_localities_table = NULL)
}
\arguments{
\item{csv_files_table}{data.frame}

\item{csv_localities_table}{data.frame}
}
\value{
data in standard format
}
\description{
This function read raw data from loggers by table saved in CSV file
}
\examples{
example_tomst_data <- microclim::mc_feed_from_csv("examples/data/TOMST/files_table.csv")
}
