% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{mc_feed_from_df}
\alias{mc_feed_from_df}
\title{Data files reading}
\usage{
mc_feed_from_df(files_table)
}
\arguments{
\item{files_table}{data.frame which describe data files}
}
\value{
data in standard format
}
\description{
This function read raw data from loggers by data.frame with files description.
Columns of data.frame:
\itemize{
\item path - path to file
\item locality_id
\item logger
\item serial_number - can be NA, than try detect
}
}
