% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{mc_feed_TMS_directory}
\alias{mc_feed_TMS_directory}
\title{Reading TMS files from directory}
\usage{
mc_feed_TMS_directory(directory, recursive = TRUE)
}
\arguments{
\item{directory}{character}

\item{recursive}{logical - recursive search in subdirectories}
}
\value{
data in standard format
}
\description{
This function read TMS data files from directory. Locality is set None.
}
\examples{
example_tms_data <- microclim::mc_feed_TMS_directory("examples/data/TMS/")
}
