% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{mc_feed_TOMST_directory}
\alias{mc_feed_TOMST_directory}
\title{Reading TOMST files from directory}
\usage{
mc_feed_TOMST_directory(directory, recursive = TRUE)
}
\arguments{
\item{directory}{character}

\item{recursive}{logical - recursive search in subdirectories}
}
\value{
data in standard format
}
\description{
This function read TOMST data files from directory. Locality is set None.
}
\examples{
example_tomst_data <- microclim::mc_feed_TOMST_directory("examples/data/TOMST/")
}
