% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{mc_feed_TOMST_files}
\alias{mc_feed_TOMST_files}
\title{Reading TOMST files}
\usage{
mc_feed_TOMST_files(files)
}
\arguments{
\item{files}{vector of character - files with data}
}
\value{
data in standard format
}
\description{
This function read data files of TOMST type. Locality is set None.
}
\examples{
example_tomst_data <- microclim::mc_feed_TOMST_files(c("examples/data/TOMST/data_91184101_0.csv", "examples/data/TOMST/data_94184102_0.csv"))
}
