% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{mc_filter}
\alias{mc_filter}
\title{Filter data}
\usage{
mc_filter(data, localities = c(), sensors = c())
}
\arguments{
\item{data}{in standard format}

\item{localities}{locality_ids for filtering data; if empty then all}

\item{sensors}{sensor_ids for filtering data; if empty then all}
}
\value{
filtered data in standard format
}
\description{
This function filter data by localities and sensors
}
\examples{
example_tomst_data1 <- mc_filter(example_tomst_data1, localities=c("A6W79", "A2E32"), sensors=c("TMS_T1", "TMS_T2"))
}
