% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_HOBODataFormat-class}
\alias{mc_HOBODataFormat-class}
\alias{mc_HOBODataFormat}
\title{Class for reading HOBO logger files}
\description{
Provides the key for reading the HOBO source files. In which column is the date,
in what format is the date, in which columns are records of which sensors.
The code defining the class is in section methods ./R/model.R
}
\seealso{
\link{mc_DataFormat}, \link{mc_data_formats}
}
