% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mc_info}
\alias{mc_info}
\title{get sensors info table}
\usage{
mc_info(data)
}
\arguments{
\item{data}{in format for preparing or calculation}
}
\value{
dataframe with columns locality_id, serial_number, sensor_id, sensor_name, start_date, end_date, step, min_value, max_value, count_values, count_na
}
\description{
This function return dataframe with info about sensors
}
