% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mc_info_clean}
\alias{mc_info_clean}
\title{Get all clean table}
\usage{
mc_info_clean(data)
}
\arguments{
\item{data}{in format for preparing}
}
\value{
dataframe with columns locality_id, serial_number, start_date, end_date, step, count_duplicits, count_missed, count_disordered
}
\description{
This function return dataframe with all clean info about loggers
}
\examples{
log_table <- mc_prep_logs(cleaned_example_tomst_data1)
}
