% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mc_info_count}
\alias{mc_info_count}
\title{count data}
\usage{
mc_info_count(data)
}
\arguments{
\item{data}{in standard format}
}
\value{
data.frame with count localities, loggers and sensors
}
\description{
This function return count of localities, loggers and sensors
}
\examples{
count_table <- mc_info_count(example_tomst_data1)
}
