% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mc_info_logger}
\alias{mc_info_logger}
\title{Get loggers info table}
\usage{
mc_info_logger(data)
}
\arguments{
\item{data}{myClim object in Raw-format. see \link{myClim-package}}
}
\value{
data.frame with columns:
\itemize{
\item locality_id - when provided by user then locality ID, when not provided identical with serial number
\item index - logger index in locality
\item serial_number - serial number of logger when provided or automatically detected from file name or header
\item logger_type - loger type
\item start_date - the oldest record on the sensor
\item end_date - the newest record on the sensor
\item step_seconds - time step of records series (seconds)
}
}
\description{
This function return data.frame with info about loggers
}
\examples{
mc_info_logger(mc_data_example_raw)
}
