% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mc_info_meta}
\alias{mc_info_meta}
\title{Get localities metadata table}
\usage{
mc_info_meta(data)
}
\arguments{
\item{data}{myClim object in Raw-format or Agg-format (see \link{myClim-package})}
}
\value{
data.frame with columns:
\itemize{
\item locality_id
\item lon_wgs84
\item lat_wgs84
\item altitude
\item tz_offset
}
}
\description{
This function return data.frame with localities metadata
}
\examples{
mc_info_meta(mc_data_example_calc)
}
