% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mc_info_range}
\alias{mc_info_range}
\title{Get table with measuring range}
\usage{
mc_info_range(data)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}
}
\value{
data.frame with columns:
\itemize{
\item sensor_name - name of sensor (e.g., TMS_T1, TMS_moist, HOBO_T) see \link{mc_data_sensors}
\item min_value - minimal value
\item max_value - maximal value
\item positive_jump - Maximal difference between two consecutive values. Next value is higher than previous.
\item negative_jump - Maximal difference between two consecutive values. Next value is lower than previous.
}
}
\description{
This function return data.frame with min value, max value and possible jumps
}
\details{
This function is useful for prepare parameter for \code{\link[=mc_states_outlier]{mc_states_outlier()}} function.
The range values are used from \code{mc_data_sensors}.
}
\examples{
mc_info_range(mc_data_example_raw)
}
