% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mc_info_states}
\alias{mc_info_states}
\title{Get states info table}
\usage{
mc_info_states(data)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}
}
\value{
data.frame with columns:
\itemize{
\item locality_id - when provided by user then locality ID, when not provided identical with serial number
\item logger_index - index of logger in locality
\item logger_type - type of logger
\item sensor_name - original sensor id if not modified, if renamed then new name (e.g.,"GDD5", "HOBO_T_mean" ,"TMS_T1_max", "my_sensor01")
\item tag - category of state
\item start - start datetime
\item end - end datetime
\item value - value of state
}
}
\description{
This function return data.frame with info about sensor states see \link{myClim-package}
}
\examples{
mc_info_states(mc_data_example_raw)
}
