% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{mc_load}
\alias{mc_load}
\title{Load myClim object}
\usage{
mc_load(file)
}
\arguments{
\item{file}{path to input .rds file}
}
\value{
loaded myClim object
}
\description{
This function loads the myClim .rds data object saved with \link{mc_save}.
The \code{mc_save} and \code{mc_load} functions secure that the myClim object is correctly
loaded across myClim versions.
}
\examples{
\dontrun{data <- mc_load("myClim_data.rds")}
}
