% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_LoggerMetadata-class}
\alias{mc_LoggerMetadata-class}
\alias{mc_LoggerMetadata}
\title{Class for logger metadata}
\description{
Class for logger metadata
}
\section{Slots}{

\describe{
\item{\code{type}}{of logger (TMS, ThermoDatalogger, Dendrometer, HOBO)}

\item{\code{serial_number}}{serial number of the logger}

\item{\code{step}}{§Time step of microclimatic data series in minutes defined by user.
In \code{\link[=mc_prep_clean]{mc_prep_clean()}} function is used instead of auto-detected value.§}
}}

