% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_MainMetadata-class}
\alias{mc_MainMetadata-class}
\alias{mc_MainMetadata}
\title{Class for myClim object metadata in Calc-format}
\description{
Class for myClim object metadata in Calc-format
}
\section{Slots}{

\describe{
\item{\code{step}}{time step of data in minutes}

\item{\code{period}}{§value from \code{\link[=mc_agg]{mc_agg()}}§}

\item{\code{intervals_start}}{§start datetime of data intervals for spacial periods all and custom (see \code{\link[=mc_agg]{mc_agg()}})§}

\item{\code{intervals_end}}{§end datetime of data intervals for spacial periods all and custom (see \code{\link[=mc_agg]{mc_agg()}})§}
}}

\seealso{
Calc-format section of \link{myClim-package}
}
