% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_MainMetadataAgg-class}
\alias{mc_MainMetadataAgg-class}
\alias{mc_MainMetadataAgg}
\title{Class for myClim object metadata in Raw-format}
\description{
Class for myClim object metadata in Raw-format
}
\section{Slots}{

\describe{
\item{\code{version}}{§of the myClim packege in which was created§}

\item{\code{format_type}}{§type of format (raw, agg)§}

\item{\code{step}}{time step of data in seconds}

\item{\code{period}}{value from \code{\link[=mc_agg]{mc_agg()}} (e.g. month, day, all...)}

\item{\code{intervals_start}}{start datetime of data intervals for spacial periods all and custom (see \code{\link[=mc_agg]{mc_agg()}})}

\item{\code{intervals_end}}{end datetime of data intervals for spacial periods all and custom (see \code{\link[=mc_agg]{mc_agg()}})}
}}

\seealso{
\link{mc_MainMetadata} \link{myClim-package}
}
