% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_OffsetCalibration-class}
\alias{mc_OffsetCalibration-class}
\alias{mc_OffsetCalibration}
\title{Class for calibration with offset}
\description{
Class for calibration with offset
}
\section{Slots}{

\describe{
\item{\code{offset}}{numeric}
}}

