% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_Physical-class}
\alias{mc_Physical-class}
\alias{mc_Physical}
\title{Class for physical}
\description{
Class for physical
}
\section{Slots}{

\describe{
\item{\code{name}}{of physical}

\item{\code{description}}{character info}

\item{\code{units}}{measurument (°C, \\%, m3/m3, raw, mm, ...)}

\item{\code{calibration_class}}{class for calibration}

\item{\code{viridis_color_map}}{viridis color map option}

\item{\code{scale_coeff}}{coefficient for plot; value * scale_coef is in range 0-1}
}}

