% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_calib}
\alias{mc_prep_calib}
\title{Sensor calibration}
\usage{
mc_prep_calib(data, sensors, localities = NULL)
}
\arguments{
\item{data}{in format for preparing or calculation}

\item{sensors}{vector of sensor names for calibration

It is not possible calibrate TMSmoisture sensor with this function.
Calibration of TMSmoisture sensor is processed in mc_calc_vwc during conversion to volumetric water content.
Only sensors with real value type can be calibrated.}

\item{localities}{vector of locality_ids, if NULL, then calibrate in all localities (default NULL)}
}
\value{
data with calibrated sensors.
}
\description{
This function calibrate values of sensor by sensor$calibration parameters. Values are changed
and parameter sensor$metadata@calibrated is set to TRUE. It isn't possible calibrate calibrated sensor again.
}
\details{

}
