% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_clean}
\alias{mc_prep_clean}
\title{Cleaning datetime series}
\usage{
mc_prep_clean(data, silent = FALSE)
}
\arguments{
\item{data}{myClim object in Prep-format (output of \code{mc_read} functions family) see e.g. \code{\link[=mc_read_files]{mc_read_files()}}}

\item{silent}{if true, then cleaning log table is not printed in console (default FALSE), see \code{\link[=mc_info_clean]{mc_info_clean()}}}
}
\value{
\itemize{
\item cleaned myClim object in Prep-format
\item cleaning log is by default printed in console, and can be called ex post by \code{\link[=mc_info_clean]{mc_info_clean()}}
}
}
\description{
Function 'mc_prep_clean' check time-series in myClim object in prep-format for missing, duplicated, and disordered records and regularize microclimatic time-series to constant time-step rounded to the nearest interval. Duplicit records are removed and missing values are filled with NA's.
See details.
}
\details{
Processing the data with \code{mc_prep_clean} is a mandatory step required for further data handling in \code{myClim} library.

This function guarantee that all time series are in chronological order and have regular time-step, without duplicated records.
\code{mc_prep_clean} identify time-step from input time series based on last 100 records. In case of irregular time series, function returns warning and skip series.

In case the time step is regular, but is not nicely rounded, function round the time series to the closest nice time and shift original data to nicely rounded time series. (e.g. original records in 10 min regular step c(11:58, 12:08, 12:18, 12:28) are shifted to newly generated nice sequence c(12:00, 12:10, 12:20, 12:30) microclimatic records are not modified but only shifted).
}
\examples{
cleaned_data <- mc_prep_clean(mc_data_example_source)
}
