% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_expandtime}
\alias{mc_prep_expandtime}
\title{Expand time steps}
\usage{
mc_prep_expandtime(
  data,
  to_step,
  localities = NULL,
  loggers = NULL,
  from_step = NULL
)
}
\arguments{
\item{data}{cleaned myClim object see \link{myClim-package}}

\item{to_step}{new time step in seconds (e.g. 3600 for one hour)}

\item{localities}{vector of locality_ids where to perform expansion, if NULL, then
expand time step on all localities (default NULL)}

\item{loggers}{vector of logger names where to perform expansion, if NULL, then
expand time step on all loggers (default NULL)}

\item{from_step}{original time step in seconds. Only loggers with this time step are expanded.
If NULL, then expand all loggers with time step longer than to_step.}
}
\value{
raw myClim data with expanded datetime.
}
\description{
This function is used to expand time step to shorter time step.
Original time step must be multiple of new time step.
The gaps are filled with NA values.
}
\examples{
mc_prep_expandtime(mc_data_example_clean, to_step = 300, localities = "A1E05", loggers = "Thermo_1")
}
