% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_fillNA}
\alias{mc_prep_fillNA}
\title{Fill NA}
\usage{
mc_prep_fillNA(
  data,
  localities = NULL,
  sensors = NULL,
  maxgap = 5,
  method = "linear"
)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{localities}{names of localities; if NULL then all (default NULL)}

\item{sensors}{names of sensors; if NULL then all (default NULL) see \code{names(mc_data_sensors)}}

\item{maxgap}{maximum number of consecutively NA values to fill (default 5)}

\item{method}{used for approximation. It is implemented now only "linear". (default "linear")}
}
\value{
myClim object with filled NA values
}
\description{
This function approximate NA (missing) values. It was designed to fill
only small gaps in microclimatic time-series therefore, the default maximum
length of the gap is 5 missing records and longer gaps are not filled
Only linear method is implemented from \link[zoo:na.approx]{zoo::na.approx} function.
}
