% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_flat}
\alias{mc_prep_flat}
\title{Flattening data}
\usage{
mc_prep_flat(data)
}
\arguments{
\item{data}{in format for preparing}
}
\value{
flattened data in format for calculation
}
\description{
This function flatten data. Logger lever from data hierarchy is deleted.
Sensors are moved to locality and datetimes are merged to one series.
}
\examples{
calc_data <- mc_prep_flat(example_cleaned_tomst_data1)
}
