% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_logs}
\alias{mc_prep_logs}
\title{Get all clean log messages}
\usage{
mc_prep_logs(data)
}
\arguments{
\item{data}{in standard format}
}
\value{
dataframe with columns locality_id, serial_number, clean_type, message
}
\description{
This function return dataframe with all clean log messages
}
\examples{
log_table <- mc_prep_logs(cleaned_example_tomst_data1)
}
