% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_merge}
\alias{mc_prep_merge}
\title{Merge myClim objects}
\usage{
mc_prep_merge(data_items)
}
\arguments{
\item{data_items}{list of myClim objects in Prep-format or Calc-format see \link{myClim-package}; Format of merged objects must be same.}
}
\value{
merged myClim object in the same format as input objects
}
\description{
This function is designed to merge myClim objects into one.
}
\details{
This function works only when the input myClim objects have the same format (Prep-format, Calc-format) and the same time step.
It is required same step in Calc-Format data. If some two data objects contains locality with same names (locality_id) in Prep-format,
than list of loggers are merged. If some two data objects contains locality with same names in Calc-format, than list os sensors are merged.
Sensors with same names are renamed.
}
\examples{
merged_data <- mc_prep_merge(list(mc_data_example_source, mc_data_example_source))
}
