% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_meta}
\alias{mc_prep_meta}
\title{Set metadata of localities}
\usage{
mc_prep_meta(data, metadata, param_name = NULL)
}
\arguments{
\item{data}{myClim object in Prep-format or Calc-formt see \link{myClim-package}}

\item{metadata}{for localities can be named list or table
\itemize{
\item named list: \code{metadata <- list(locality_id=value)}; \code{param_name} must be set
\item table with column \code{locality_id} and another columns named by metadata parameter name; \code{param_name} must be NULL
}}

\item{param_name}{name of locality metadata parameter; Default names are \code{altitude}, \code{lat_wgs84}, \code{lon_wgs84}, \code{tz_offset}.
Another names are inserted to \code{user_data} list. see \link{mc_LocalityMetadata}}
}
\value{
MyClim object in the same format as input, with locality metadata filled in
}
\description{
MUST BE CHANGED - is from mc_prep_user_tz
This function allow user to set the offset in minutes against original time series of microclimatic records.
\code{MyClim} generally assume the records are in UTM Time Zone. When offset is provided, myClim functions by default use
the time corrected with this offset in calculations. See details.
}
\details{
If is set tz_offset value, than tz_type is set to \verb{user defined} value.

MUST BE CHANGED - is from mc_prep_user_tz
For analysis of microclimatic data it is important to think of time zones because of diurnal or seasonal microclimatic rhythms.
\code{mc_prep_user_tz} allow user to set time offset for individual localities to harmonize e.g.day/night cycles across vhole data set.

This function can be used also inversely, for heterogeneous data sets containing loggers recording in local time
and user wish to unify them by setting individual offest e.g. to UTM. This function is also useful for corrections
of shifted loggers time series e.g. due to technical issue.

In case user is sure, the loggers recorded in UTC and wants to harmonize data set to solar time,
there is \code{\link[=mc_prep_solar_tz]{mc_prep_solar_tz()}} calculating offset based on coordinates to harmonize the midday across vhole dataset.
}
\examples{
data <- mc_prep_meta(mc_data_example_source, list(`91184101`=60), param_name="tz_offset")
}
