% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_meta_sensor}
\alias{mc_prep_meta_sensor}
\title{§Set metadata of sensors}
\usage{
mc_prep_meta_sensor(
  data,
  values,
  param_name,
  localities = NULL,
  logger_types = NULL
)
}
\arguments{
\item{data}{myClim onject in Prep-format or Calc-format see \link{myClim-package}}

\item{values}{named list with metadata values; names of items are sensor_names e.g. change height \code{list(TMS_T1="soil 8 cm")}}

\item{param_name}{name of locality metadata parameter; Possible names are \code{name} and \code{height}.}

\item{localities}{vector of \code{locality_id} values where to change metadata; if NULL than all localities (default NULL)}

\item{logger_types}{vector of \code{logger_type} values where to change metadata; if NULL than all loggers (default NULL);
parameter is useful only for Prep-format of myClim having the level of logger see \link{myClim-package}}
}
\value{
myClim object in the same format as input, with sensor metadata filled in§
}
\description{
This function allows you to modify sensor metadata. Function can rename sensors too. See \link{mc_SensorMetadata}
}
\examples{
data <- mc_prep_meta_sensor(mc_data_example_source, list(TMS_T1="my_TMS_T1"), param_name="name")
}
