% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_rename_locality}
\alias{mc_prep_rename_locality}
\title{rename locality_id}
\usage{
mc_prep_rename_locality(data, locality_ids)
}
\arguments{
\item{data}{myClim object in Prep-format or Calc-format see \link{myClim-package}}

\item{locality_ids}{list with new locality_ids; names of items are old ones e.g. \code{list(A1E05="ABC05", A2E32="CDE32")}}
}
\value{
The same myClim object as input but with changed locality_ids
}
\description{
This function change locality_ids.
}
\examples{
data <- mc_prep_rename_locality(example_tomst_data1, list(A1E05="ABC05", A2E32="CDE32"))
}
