% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_rename_sensor}
\alias{mc_prep_rename_sensor}
\title{Rename sensor}
\usage{
mc_prep_rename_sensor(
  data,
  sensor_names,
  localities = NULL,
  serial_numbers = NULL
)
}
\arguments{
\item{data}{myClim onject in Prep-format or Calc-format see \link{myClim-package}}

\item{sensor_names}{named list with new names of sensors; names of items are old ones e.g. \code{list(TMS_T1="TMS_Tsoil")}== \code{list("Old"="New")}}

\item{localities}{vector of locality_ids where to rename sensors; if NULL than all localities (default NULL)}

\item{serial_numbers}{vector of serial_numbers of loggers; if NULL than all loggers (default NULL); parameter is useful only for
Prep-format of myClim having the level of logger see \link{myClim-package}}
}
\value{
the same myClim object as input with changed sensor names
}
\description{
This function rename sensors.
}
\examples{
load("data/mc_data_example_calc.rda")
data <- mc_prep_rename_sensor(c_data_example_calc, list(TMS_T1="TMS_Tsoil"))
}
