% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_solar_tz}
\alias{mc_prep_solar_tz}
\title{Solar TZ offset}
\usage{
mc_prep_solar_tz(data)
}
\arguments{
\item{data}{in format for preparing or calculation}
}
\value{
data with changed TZ offset in same format as input data
}
\description{
This function compute TZ offset in localities by solar time
}
\details{
The function require filled longitude of locality in slot lon_wgs84 of metadata.
TZ offset in minutes is calculated as \code{longitude / 180 * 12 * 60}.
}
\examples{
cleaned_example_tomst_data1 <- mc_prep_solar_tz(cleaned_example_tomst_data1)
}
