% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_TMSoffsoil}
\alias{mc_prep_TMSoffsoil}
\title{TMS Logger Data Anomaly Detection}
\usage{
mc_prep_TMSoffsoil(
  data,
  localities = NULL,
  soil_sensor = mc_const_SENSOR_TMS_T1,
  air_sensor = mc_const_SENSOR_TMS_T2,
  moist_sensor = mc_const_SENSOR_TMS_moist,
  output_sensor = "off_soil",
  smooth = FALSE,
  smooth_window = 10,
  smooth_threshold = 0.5,
  sd_threshold = 0.76085,
  minmoist_threshold = 721.5
)
}
\arguments{
\item{data}{cleaned myClim object see \link{myClim-package}}

\item{localities}{names of localities; if NULL then all (default NULL)}

\item{soil_sensor}{character, soil temperature sensor (default \code{mc_const_SENSOR_TMS_T1})}

\item{air_sensor}{character, air temperature sensor (default \code{mc_const_SENSOR_TMS_T2})}

\item{moist_sensor}{character, soil moisture sensor (default \code{mc_const_SENSOR_TMS_moist})}

\item{output_sensor}{(default "off_soil")}

\item{smooth}{logical, smooth out isolated faulty/correct records using floating window (default FALSE)}

\item{smooth_window}{integer, smooth floating window width (in days) (default 10)}

\item{smooth_threshold}{numeric, floating window threshold for detection of faulty records. Defaults to "0.5")}

\item{sd_threshold}{numeric, threshold value for the ratio of the standard deviation of the soil sensor
to the above-ground sensor temperatures (default 0.76085)}

\item{minmoist_threshold}{numeric, threshold value for the minimum soil moisture (default 721.5)}
}
\value{
numeric vector (0 = correct measurement, 1 = faulty measurement) stored as virtual sensor in myClim object
}
\description{
This function creates new virtual sensor labelling anomalies in TMS logger caused by displacement from soil.
}
\details{
TMS loggers, when correctly installed in the soil, exhibit certain temperature and soil moisture signal characteristics.
Temperature varies most significantly at the soil interface, and temperature fluctuations in the soil are minimized.
The moisture signal from a sensor that has lost direct contact with the soil is reduced.
The following criteria are used for detecting faulty measurements: the ratio of the standard deviation of the soil
sensor to the above-ground sensor is greater than the defined threshold (default 0.76085), and simultaneously, the minimum soil moisture is less than 721.5.
The calculation of variables used for classification is performed in a centered floating window with a width of one day.
Optionally, the prediction results can be smoothed using a floating window.
Selection and parametrization of criteria was done using a recursive partitioning (rpart::rpart)
on the set of 154 TMS timeseries from different environmental settings (total 7.8M records, measurements from temperate forest, tropical rainforest, alpine zone)
with manually labelled erroneos records.  Note that sensitivity (true positive rate) on this training dataset was 95.1\% and specificity (true negative rate) 99.4\%.
Smoothing with 10 day floating window increased sensitivity to 96.8\% while retaining specifity at the same level.
}
\examples{
data <- mc_prep_TMSoffsoil(mc_data_example_clean)
}
