% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_trim}
\alias{mc_prep_trim}
\title{Trim records from beginning and/or end}
\usage{
mc_prep_trim(data, start = TRUE, end = FALSE, n = 1, logger_type = NULL)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{start}{logical; trim first \code{n} records (default TRUE)}

\item{end}{logical; trim last \code{n} records (default FALSE)}

\item{n}{number of records to trim from selected side(s); must be positive integer (default 1)}

\item{logger_type}{optional filter; vector of logger types to trim in Raw-format (default NULL)}
}
\value{
myClim object in the same format as input, with trimmed time-series.
}
\description{
This function removes the first and/or last \code{n} records from each time-series.
It works for both Raw and Agg formats.
}
\details{
In Raw-format, trimming can be restricted to selected logger types with \code{logger_type}.
In Agg-format, \code{logger_type} is ignored.

If both \code{start} and \code{end} are TRUE, the function trims \code{n} records from both sides.
When the resulting number of records is negative, the output time-series is empty.
}
\examples{
data_trim <- mc_prep_trim(mc_data_example_clean, start=TRUE, end=TRUE,
                          n=1, logger_type=c("Thermo", "TMS"))
}
