% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_user_tz}
\alias{mc_prep_user_tz}
\title{Set user defined offset against original time}
\usage{
mc_prep_user_tz(data, tz_offsets)
}
\arguments{
\item{data}{myClim object in raw, clean, or calculation format}

\item{tz_offsets}{named list (name: \code{locality_id}, list item: \code{tz_offset} in rounded minutes)}
}
\value{
MyClim object in the same format as input, with \code{tz_offset} filled in locality metadata
}
\description{
This function allow user to set the offset in minutes against original time series of microclimatic records. \code{MyClim} generally assume the records are in UTM Time Zone. When offset is provided, myClim functions by default use the time corrected with this offset in calculations. See details.
}
\details{
For analysis of microclimatic data it is important to think of time zones because of diurnal or seasonal microclimatic rhythms. \code{mc_prep_user_tz} allow user to set time offset for individual localities to harmonize e.g.day/night cycles across vhole data set.

This function can be used also inversely, for heterogeneous data sets containing loggers recording in local time and user wish to unify them by setting individual offest e.g. to UTM. This function is also useful for corrections of shifted loggers time series e.g. due to technical issue.

In case user is sure, the loggers recorded in UTC and wants to harmonize data set to solar time, there is \code{\link[=mc_prep_solar_tz]{mc_prep_solar_tz()}} calculating offset based on coordinates to harmonize the midday across vhole dataset.
}
\examples{
example_tomst_data2 <- mc_prep_solar_tz(example_tomst_data2, list(`91184101`=60))
}
