% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{mc_read_data}
\alias{mc_read_data}
\title{Reading files with locality metadata}
\usage{
mc_read_data(files_table, localities_table = NULL)
}
\arguments{
\item{files_table}{path to csv file or data.frame object containing 4 columns:
\itemize{
\item path - path to file
\item locality_id
\item data_format
\item serial_number - can be NA, than try detect
}}

\item{localities_table}{path to csv file or data.frame object containing 5 columns:
\itemize{
\item locality_id
\item altitude
\item lon_wgs84
\item lat_wgs84
\item tz_offset
}}
}
\value{
myClim object in Prep-format see \link{myClim-package}
}
\description{
This function requires two tables. i.) \code{files_table} with paths pointing to raw
csv logger files, specification of data format (logger type) and locality name.
ii) \code{localities_table} with locality id and metadata e.g. longitude, latitude, altitude...
}
\details{
The input tables could be R data.frames or csv files. When loading \code{files_table} and \code{localities_table} from external CSV it
must have header, column separator must be comma ","
}
\examples{
\dontrun{
tomst_data <- mc_read_data("examples/data/TOMST/files_table.csv", "examples/data/TOMST/localities_table.csv")
}
}
