% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{mc_read_directory}
\alias{mc_read_directory}
\title{Reading files from directory}
\usage{
mc_read_directory(directory, dataformat_name, recursive = TRUE)
}
\arguments{
\item{directory}{character}

\item{dataformat_name}{character - data format of logger (TOMST, TOMST_join)}

\item{recursive}{logical - recursive search in subdirectories}
}
\value{
data in standard format
}
\description{
This function read csv data files from directory of one logger type.
If csv file is not in correct format, is skipped. Locality is set to serial_number of logger.
}
\examples{
example_tomst_data <- myClim::mc_read_directory("examples/data/TOMST/", "TOMST")
}
