% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{mc_read_files}
\alias{mc_read_files}
\title{Reading files}
\usage{
mc_read_files(files, dataformat_name)
}
\arguments{
\item{files}{vector of character - files with data}

\item{dataformat_name}{character - data format of logger (TOMST, TOMST_join)}
}
\value{
data in standard format
}
\description{
This function read data files of one logger type. Locality is set to serial_number of logger.
}
\examples{
example_tomst_data <- myClim::mc_read_files(c("examples/data/TOMST/data_91184101_0.csv", "examples/data/TOMST/data_94184102_0.csv"), "TOMST")
}
