% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{mc_read_long}
\alias{mc_read_long}
\title{Reading data from long data.frame}
\usage{
mc_read_long(data_table, sensor_ids = list(), clean = TRUE, silent = FALSE)
}
\arguments{
\item{data_table}{long data.frame with Columns:
\itemize{
\item locality_id
\item sensor_name  - see \code{names(mc_data_sensors)}
\item datetime - POSIXct and UTC timezone is required
\item value
}}

\item{sensor_ids}{list with relations between sensor_names and sensor_ids (default list());
sensor_id is key from \code{names(mc_data_sensors)}. If sensor_name is the same as sensor_id,
it does not have to be in the list.

\code{sensor_ids <- list(sensor_name1=sensor_id1, sensor_name2=sensor_id2)}}

\item{clean}{§if TRUE, then \link{mc_prep_clean} is called automatically (default TRUE)§}

\item{silent}{§if TRUE, then any information is printed in console (default FALSE)§}
}
\value{
myClim object in Prep-format
}
\description{
This is universal function designed to read time series and values
from long data.frame to myClim object.
}
\details{
Similar like \link{mc_read_wide} but is capable to read multiple sensors.
}
\seealso{
\link{mc_read_wide}
}
