% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\docType{data}
\name{mc_read_problems}
\alias{mc_read_problems}
\title{Environment for reading problems}
\format{
An object of class \code{environment} of length 0.
}
\usage{
mc_read_problems
}
\description{
Environment for reading problems
}
\keyword{datasets}
