% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{mc_read_table_wide}
\alias{mc_read_table_wide}
\title{Reading universal data from wide data.frame}
\usage{
mc_read_table_wide(
  data_table,
  sensor_id = myClim:::.model_const_SENSOR_real,
  sensor_name = NULL
)
}
\arguments{
\item{data_table}{data.frame with values

Columns:
\itemize{
\item datetime column - POSIXct and UTC timezone is required
\item Name of locality\link{1} - values
\item ...
\item Name of locality\link{n} - values
}}

\item{sensor_id}{name of sensor from \link{mc_data_sensors} (default real)}

\item{sensor_name}{custom name of sensor; if NULL than \code{sensor_name <- sensor_id} (default NULL)}
}
\value{
data in prep-format
}
\description{
This function read data from data.frame. There is datetime column. Other columns mean locality values.
Name of column mean name of locality. All values are from one sensor.
}
