% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{mc_read_wide}
\alias{mc_read_wide}
\title{Reading data from wide data.frame}
\usage{
mc_read_wide(
  data_table,
  sensor_id = myClim:::.model_const_SENSOR_real,
  sensor_name = NULL
)
}
\arguments{
\item{data_table}{data.frame with first column of POSIXct time format UTC timezone,
followed by columns with microclimatic records. See details.

Columns:
\itemize{
\item datetime column - POSIXct and UTC timezone is required
\item Name of locality\link{1} - values
\item ...
\item Name of locality\link{n} - values
}}

\item{sensor_id}{define the sensor type, one of \link{mc_data_sensors} (default \code{real})}

\item{sensor_name}{custom name of sensor; if NULL than \code{sensor_name <- sensor_id} (default NULL)}
}
\value{
myClim object in Prep-format
}
\description{
This is universal function designed to read time series and values
from wide data.frame to myClim object.
}
\details{
The first column of input data.frame must be datetime column in POSIXct time format UTC timezone.
Following columns represents localities. Column names are the localities.
All values in wide data.frame represents the same sensor, e.g. air temperature. If you wish to
read multiple sensors use \link{mc_read_long} or use \link{mc_read_wide} multiple times separately
for each sensor and that merge myClim objects with \link{mc_prep_merge}
}
\seealso{
\link{mc_read_long}
}
