% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{mc_reshape_long}
\alias{mc_reshape_long}
\title{Longformat of sensor values}
\usage{
mc_reshape_long(data, localities = NULL, sensors = NULL)
}
\arguments{
\item{data}{in format for preparing or calculation}

\item{localities}{locality_ids; if NULL then all (default NULL)}

\item{sensors}{names of sensors; if NULL then all (default NULL)}
}
\value{
data.frame with columns locality_id, serial_number, sensor, datetime, value
}
\description{
This function create data.frame with values of sensor
}
\examples{
example_tms_t1_table <- myClim::mc_reshape_long(example_tomst_data, c("A6W79", "A2E32"), c("TMS_T1", "TMS_T2"))
}
