% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{mc_reshape_longformat}
\alias{mc_reshape_longformat}
\title{Longformat of sensor values}
\usage{
mc_reshape_longformat(data, localities = c(), sensors = c())
}
\arguments{
\item{data}{all data in standard format}

\item{localities}{names of localities; if empty then all}

\item{sensors}{names of sensors; if empty then all}
}
\value{
data.frame with columns location, serial_number, sensor, datetime, value
}
\description{
This function create data.frame with values of sensor
}
\examples{
example_tms_t1_table <- microclim::mc_reshape_longformat(example_tomst_data, c("A6W79", "A2E32"), c("TMS_T1", "TMS_T2"))
}
