% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{mc_reshape_wide}
\alias{mc_reshape_wide}
\title{Wideformat of sensor values}
\usage{
mc_reshape_wide(data, localities = c(), sensors = c())
}
\arguments{
\item{data}{all data in standard format}

\item{localities}{names of localities; if empty then all}

\item{sensors}{names of sensors; if empty then all}
}
\value{
data in standard format
}
\description{
This function create data.frame with values of sensor in wide format.
}
\examples{
example_tms_wideformat <- mc_reshape_wide(example_tomst_data1, c("A6W79", "A2E32"), c("TMS_T1", "TMS_T2"))
}
