% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{mc_reshape_wideformat}
\alias{mc_reshape_wideformat}
\title{Sensor values by localities}
\usage{
mc_reshape_wideformat(data, sensor, localities)
}
\arguments{
\item{data}{all data in standard format}

\item{sensor}{name of sensor}

\item{localities}{names of localities}
}
\value{
data in standard format
}
\description{
This function create data.frame with values of sensor
}
\examples{
example_tms_t1_table <- microclim::mc_reshape_wideformat(example_tms_data, "T1", c("LOC_1", "LOC_2"))
}
