% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{mc_reshape_wideformat_interval}
\alias{mc_reshape_wideformat_interval}
\title{Wideformat of sensor values by interval}
\usage{
mc_reshape_wideformat_interval(
  data,
  localities = c(),
  sensors = c(),
  interval_length = 15
)
}
\arguments{
\item{data}{all data in standard format}

\item{localities}{names of localities; if empty then all}

\item{sensors}{names of sensors; if empty then all}

\item{interval_length}{in minutes (default 15)}
}
\value{
data in standard format
}
\description{
This function create data.frame with values of sensor in wide format.
Mean is computed from values in datetime interval.
}
\examples{
example_tms_wideformat_interval <- microclim::mc_reshape_wideformat_interval(example_tms_data, c("LOC_1", "LOC_2"), c("T1", "T2"), 10)
}
