% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{mc_save}
\alias{mc_save}
\title{Save myClim object}
\usage{
mc_save(data, file)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{file}{path to output RDS file}
}
\description{
§This function save the myClim data object to an RDS file. The standard R functions save and saveRDS are unsafe
to save the MyClim object. The mc_save a mc_load functions ensure that the myClim object is correctly loaded
in the newer version.§
}
