% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{mc_save_localities}
\alias{mc_save_localities}
\title{Save myClim object separated by localities}
\usage{
mc_save_localities(data, directory)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{directory}{path to output directory}
}
\value{
RDS files saved at the output path destination
}
\description{
This function was designed for saving the myClim data object to multiple
.rds files, which every contains data of one locality. Every file is named by \code{locality_id}.
}
\examples{
tmp_dir <- tempdir()
tmp_dir <- file.path(tmp_dir, "localities")
dir.create(tmp_dir)
mc_save_localities(mc_data_example_agg, tmp_dir)
unlink(tmp_dir, recursive = TRUE)
}
