% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_Sensor-class}
\alias{mc_Sensor-class}
\alias{mc_Sensor}
\title{Class for sensor definition}
\description{
Class for sensor definition
}
\section{Slots}{

\describe{
\item{\code{sensor_id}}{of sensor (TMS_T1, TMS_T2, TMS_T3, TMS_moisture, ...)}

\item{\code{logger}}{name of logger (TMS, ...)}

\item{\code{physical}}{measurement (T, TMS_moisture, ...)}

\item{\code{default_height}}{default height of sensor in m}

\item{\code{min_value}}{minimal value}

\item{\code{max_value}}{maximal value}
}}

