% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_SensorMetadata-class}
\alias{mc_SensorMetadata-class}
\alias{mc_SensorMetadata}
\title{Class for sensor metadata}
\description{
Class for sensor metadata
}
\section{Slots}{

\describe{
\item{\code{sensor_id}}{character}

\item{\code{name}}{character}

\item{\code{height}}{in meters}

\item{\code{calibrated}}{logical - detect if sensor is calibrated}
}}

