% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{mc_states_from_sensor}
\alias{mc_states_from_sensor}
\title{Convert sensor to state}
\usage{
mc_states_from_sensor(
  data,
  source_sensor,
  tag,
  to_sensor,
  value = NA,
  inverse = FALSE
)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{source_sensor}{logical sensor for convert to states.}

\item{tag}{tag of new states.}

\item{to_sensor}{vector of sensor names where states are stored.}

\item{value}{value of new states (default NA)}

\item{inverse}{if FALSE states are created for TRUE periods of \code{source_sensor} (default FALSE)}
}
\value{
myClim object in the same format as input, with added states
}
\description{
This function create state from source sensor.
}
\examples{
data <- mc_calc_snow(mc_data_example_agg, "TMS_T2", output_sensor="snow")
data <- mc_states_from_sensor(data, source_sensor="snow", tag="snow", to_sensor="TMS_T2")
}
