% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{mc_states_to_sensor}
\alias{mc_states_to_sensor}
\title{Convert states to logical sensor}
\usage{
mc_states_to_sensor(
  data,
  tag,
  to_sensor,
  source_sensor = NULL,
  inverse = FALSE
)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{tag}{tag of states.}

\item{to_sensor}{output vector of sensor names.

\if{html}{\out{<div class="sourceCode">}}\preformatted{ If `to_sensor` is a single sensor name, then the logical sensor is created
 from union of states of all sensors with the same tag. If `to_sensor` is multiple sensor names,
 the length of the vector must be the same as length of `source_sensor`.
}\if{html}{\out{</div>}}}

\item{source_sensor}{sensors with states. If NULL, states from all sensors are used. (default NULL)}

\item{inverse}{if TRUE the value of sensor is FALSE for state intervals (default FALSE)}
}
\value{
myClim object in the same format as input, with added sensors
}
\description{
This function create logical sensor from states.
}
\examples{
states <- data.frame(locality_id="A1E05", logger_index=1, sensor_name="Thermo_T", tag="error",
                     start=lubridate::ymd_hm("2020-10-28 9:00"),
                     end=lubridate::ymd_hm("2020-10-28 9:30"))
data <- mc_states_insert(mc_data_example_clean, states)
data <- mc_states_to_sensor(data, tag="error", to_sensor="error_sensor")
}
