% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{mc_states_update}
\alias{mc_states_update}
\title{Update states of sensors}
\usage{
mc_states_update(data, states_table)
}
\arguments{
\item{data}{cleaned myClim object see \link{myClim-package}}

\item{states_table}{data.frame with columns:
\itemize{
\item locality_id
\item logger_index - index of logger in locality
\item sensor_name - original sensor id if not modified, if renamed then new name (e.g.,"GDD5", "HOBO_T_mean" ,"TMS_T1_max", "my_sensor01")
\item tag - category of state
\item start - start datetime
\item end - end datetime
\item value - value of state
}

As input can be used result of \code{\link[=mc_info_states]{mc_info_states()}}. All columns are required.}
}
\value{
myClim object in the same format as input, with updated sensor states
}
\description{
This function update states table in sensors see \link{myClim-package}. Previous states are replaced by states from table.

The date times start and end are cropped by logger/locality datetime range andthey are rounded by step.
}
\examples{
states <- mc_info_states(mc_data_example_clean)
states$value <- basename(states$value)
data <- mc_states_update(mc_data_example_clean, states)
}
