% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_TMSDataFormat-class}
\alias{mc_TMSDataFormat-class}
\alias{mc_TMSDataFormat}
\title{Class for source file data format for TMS logger}
\description{
Class for source file data format for TMS logger
}
