% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_TMSmoistureCalibration-class}
\alias{mc_TMSmoistureCalibration-class}
\alias{mc_TMSmoistureCalibration}
\title{Class for calibration of TMSmoisture sensor}
\description{
Class for calibration of TMSmoisture sensor
}
\section{Slots}{

\describe{
\item{\code{intercept}}{numeric}

\item{\code{slope}}{numeric}
}}

