% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_TOMSTJoinDataFormat-class}
\alias{mc_TOMSTJoinDataFormat-class}
\alias{mc_TOMSTJoinDataFormat}
\title{Class for source file data format for joined TOMST logger}
\description{
Class for source file data format for joined TOMST logger
}
