% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclim-package.R
\docType{package}
\name{microclim-package}
\alias{microclim}
\alias{microclim-package}
\title{microclim: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
