% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{model.DataFormat-class}
\alias{model.DataFormat-class}
\alias{model.DataFormat}
\title{Class for source file data format}
\description{
Class for source file data format
}
\section{Slots}{

\describe{
\item{\code{has_header}}{columns separator}

\item{\code{separator}}{columns separator}

\item{\code{date_column}}{index of date column}

\item{\code{date_format}}{format of date}

\item{\code{na_strings}}{strings for NA values}

\item{\code{columns}}{list with names and indexes of value columns}

\item{\code{filename_serial_number_pattern}}{character pattern for detecting serial_number from filename}

\item{\code{data_row_pattern}}{character pattern for detecting right file format}
}}

