% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{model.Sensor-class}
\alias{model.Sensor-class}
\title{Class for sensor definition}
\description{
Class for sensor definition
}
\section{Slots}{

\describe{
\item{\code{name}}{name of sensor (T1, T2, T3, TDT, RH, ...)}

\item{\code{logger}}{name of logger (TMS, TMS-T1, TMS Dendro, iButton Hygrochron, iButton Thermochron, HOBO RH, HOBO T, ...)}

\item{\code{physical}}{measurement (T, RH, VWC, ...)}

\item{\code{units}}{measurument (°C, %, m3/m3, raw, mm, ...)}

\item{\code{default_height}}{default height of sensor in cm}

\item{\code{min_value}}{minimal value}

\item{\code{max_value}}{maximal value}
}}

