% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{model.TMSDataFormat-class}
\alias{model.TMSDataFormat-class}
\alias{model.TMSDataFormat}
\title{Class for source file data format for TMS logger}
\description{
Class for source file data format for TMS logger
}
