% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{myClimList}
\alias{myClimList}
\title{Custom list for myClim object §§§}
\usage{
myClimList(metadata = NULL, localities = list())
}
\arguments{
\item{metadata}{of data object}

\item{localities}{list of licalities}
}
\description{
Top level list for store myClim data. (see \link{myClim-package})
}
