% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare.read}
\alias{prepare.read}
\title{Raw data reading}
\usage{
prepare.read(filename, serial_number, id_locality, data_format, tz = "UTC")
}
\arguments{
\item{filename}{path to source file}

\item{serial_number}{serial number of logger}

\item{id_locality}{id of locality}

\item{data_format}{definition of format source file}

\item{tz}{time zone of dates in source data}
}
\value{
data in standard format
}
\description{
This function read raw data of loger and return it in srdard format.
}
