% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare.read_files}
\alias{prepare.read_files}
\title{Data files reading}
\usage{
prepare.read_files(files_table)
}
\arguments{
\item{files_table}{data.frame}
}
\value{
data in standard format
}
\description{
This function read raw data from loggers by data.frame with files description
}
