% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare.read_files_by_csv}
\alias{prepare.read_files_by_csv}
\title{Data files reading by CSV}
\usage{
prepare.read_files_by_csv(csv_with_files_table)
}
\arguments{
\item{csv_with_files_table}{data.frame}
}
\value{
data in standard format
}
\description{
This function read raw data from loggers by table saved in CSV file
}
