% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare.read_files_by_table}
\alias{prepare.read_files_by_table}
\title{Data files reading}
\usage{
prepare.read_files_by_table(files_table)
}
\arguments{
\item{files_table}{data.frame which describe data files
- columns: path, locality_id, logger, serial_number - can be NA, than try detect}
}
\value{
data in standard format
}
\description{
This function read raw data from loggers by data.frame with files description
}
