% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare.read_TMS1_logger}
\alias{prepare.read_TMS1_logger}
\title{TMS1 logger raw data reading}
\usage{
prepare.read_TMS1_logger(filename, serial_number = NULL)
}
\arguments{
\item{filename}{path to source file}

\item{serial_number}{serial number of logger, if is NULL try detect from filename}
}
\value{
data in standard format
}
\description{
This function read raw data of TMS1 logger and return data in standard format.
}
