% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare.read_TMS3_logger}
\alias{prepare.read_TMS3_logger}
\title{TMS3 logger raw data reading}
\usage{
prepare.read_TMS3_logger(filename, serial_number, id_locality)
}
\arguments{
\item{filename}{path to source file}

\item{serial_number}{serial number of logger}

\item{id_locality}{id of locality}
}
\value{
data in standard format
}
\description{
This function read raw data of TMS3 logger and return data in standard format.
}
