% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare.read_TMS_directory}
\alias{prepare.read_TMS_directory}
\title{Reading TMS files from directory}
\usage{
prepare.read_TMS_directory(directory, recursive = TRUE)
}
\arguments{
\item{directory}{character}

\item{recursive}{logical - recursive search in subdirectories}
}
\value{
data in standard format
}
\description{
This function read TMS data files from directory. Locality is set None.
}
