% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.myClimList}
\alias{print.myClimList}
\title{Print function for myClim object}
\usage{
\method{print}{myClimList}(x, ...)
}
\arguments{
\item{x}{myClim object see \link{myClim-package}}
}
\description{
Function print info about first localities.
}
