% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.get_sensor_values_from_localities}
\alias{read.get_sensor_values_from_localities}
\title{Sensor values by localities}
\usage{
read.get_sensor_values_from_localities(data, sensor, localities)
}
\arguments{
\item{data}{all data in standard format}

\item{sensor}{name of sensor}

\item{localities}{names of localities}
}
\value{
data in standard format
}
\description{
This function create data.frame with values of sensor
}
\examples{
example_tms_t1_table <- microclim::read.get_sensor_values_from_localities(example_tms_data, "T1", c("LOC_1", "LOC_2"))
}
