% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{Sensor-class}
\alias{Sensor-class}
\title{Class for sensor definition}
\description{
Class for sensor definition
}
\section{Slots}{

\describe{
\item{\code{name}}{of sensor (T1, T2, T3, TDT, RH, ...)}

\item{\code{logger}}{name of logger (TMS3, TMS4 long, TMS-T1, TMS Dendro, iButton Hygrochron, iButton Thermochron, HOBO RH, HOBO T, ...)}

\item{\code{physical}}{measurement (T, RH, VWC, ...)}

\item{\code{units}}{measurument (°C, %, m3/m3, raw, mm, ...)}

\item{\code{defaultHeight}}{default height of sensor in cm}

\item{\code{minValue}}{minimal value}

\item{\code{maxValue}}{maximal value}
}}

