% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mc_calc_fdd}
\alias{mc_calc_fdd}
\title{Freezing Degree Days}
\usage{
mc_calc_fdd(data, sensor, output_prefix = "FDD", t_base = 0, localities = NULL)
}
\arguments{
\item{data}{myClim object in cleaned Prep-format or Calc-formt see \code{\link[=mc_agg]{mc_agg()}} and \link{myClim-package}}

\item{sensor}{name of temperature sensor used fot FDD calculation e.g. TMS_T3 see \code{names(mc_data_sensors)}}

\item{output_prefix}{name prefix of new FDD sensor (default "FDD")

name of output sensor consists of output_prefix and value t_base}

\item{t_base}{threshold temperaturefor calculation FDD (default 0)}

\item{localities}{list of locality_ids for calculation; if NULL then all (default NULL)}
}
\value{
The same myClim object as input but with added GDD sensor
}
\description{
Function add new virtual sensor with values of FDD Freezing Degree Days.
}
\details{
Maximal allowed step length for FDD calculation is day and shorter. Function creates new virtual sensor with
the same time step as input data. I. e. when the time step is shorter than a day than freezing degree day
is divided into smaller time step but still summing the day. For shorter intervals than the day the FDD value
is the contribution of the interval to the freezing degree day.
Be careful while aggregating freezing degree days to longer periods see \code{\link[=mc_agg]{mc_agg()}} only meaningful
aggregation function is \code{sum}, but user is allowed to apply anything.
}
\examples{
fdd_data <- mc_calc_fdd(mc_data_example_calc, "TMS_T3", localities = c("A2E32", "A6W79"))
fdd_agg <- mc_agg(fdd_data, list(TMS_T3=c("min", "max"), FDD5="sum"), period="day")
}
