% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mc_calc_gdd}
\alias{mc_calc_gdd}
\title{Growing Degree Days}
\usage{
mc_calc_gdd(data, sensor, output_prefix = "GDD", t_base = 5, localities = NULL)
}
\arguments{
\item{data}{in format for calculation}

\item{sensor}{name of temperature sensor}

\item{output_prefix}{name prefix of new GDD sensor (default "GDD")

name of output sensor consists of output_prefix and value t_base}

\item{t_base}{threshold temperaturefor calculation GDD (default 5)}

\item{localities}{list of locality_ids for calculation; if NULL then all (default NULL)}
}
\value{
input data with added GDD sensor
}
\description{
Function add new virtual sensor with values of GDD Growing Degree Days.
}
\details{
Maximal step length of data is day.
}
