% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mc_calc_snow}
\alias{mc_calc_snow}
\title{Snow detection from temperature}
\usage{
mc_calc_snow(
  data,
  sensor,
  output_sensor = "snow",
  localities = NULL,
  range = 2,
  tmax = 0.5,
  days = 1
)
}
\arguments{
\item{data}{myClim object in Calc-format. See \code{\link[=mc_agg]{mc_agg()}} and \link{myClim-package}}

\item{sensor}{name of temperature sensor used for snow estimation. (e.g. TMS_T2)}

\item{output_sensor}{name of output snow sensor (default "snow")}

\item{localities}{list of locality_ids where snow sill be calculated; if NULL then all (default NULL)}

\item{range}{maximal temperature range threshold for snow-covered sensor}

\item{tmax}{maximal temperature threshold for snow-covered sensor}

\item{days}{number of days to be used for moving-window for snow detection algorithm}
}
\value{
myClim object with added virtual sensor 'snow' (logical) indicating snow presence.
}
\description{
Function detects snow cover from temperature time-series.
}
\details{
Function detects snow cover from temperature time-series. Temperature sensor is considered as covered by snow when the maximal temperature in the preceding or subsequent time-window (specified by 'days' param) does not exceed specific 'tmax' threshold value (default 0.5°C) and the temperature range remain below specified 'range' threshold (default 2°C). This function rely on insulating effect of snow, significantly reducing diurnal temperature variation. Typically, for snow detection is used temperature sensor near the ground (e.g. 'TMS_T2').
The function returns vector of snow cover with same time-step as input data. To get number of days with snow cover and more info, apply \link{mc_calc_snow_agg}.
}
\examples{
snow <- mc_calc_snow(example_tomst_data1, "TMS_T2", output_sensor="TMS_T2_snow")
}
