% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mc_calc_snow_agg}
\alias{mc_calc_snow_agg}
\title{Snow detection summary}
\usage{
mc_calc_snow_agg(data, snow_sensor, localities = NULL, period = 3, use_utc = F)
}
\arguments{
\item{data}{in format for calculation}

\item{snow_sensor}{name of snow sensor created by function mc_calc_snow}

\item{localities}{names of localities; if empty then all}

\item{period}{count days for continuous cover of snow (default 3)}

\item{use_utc}{if set FALSE then datetime changed by locality tz_offset; default FALSE}
}
\value{
data.frame with columns locality, snow_days, first_day, last_day, first_day_period, last_day_period
}
\description{
Function return summary info about snow detection
}
\details{
If snow_sensor isn't in locality, then NA returned.
}
\examples{
snow_agg <- mc_calc_snow_agg(example_tomst_data1, "TMS_T3")
}
