% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{mc_calc_vwc}
\alias{mc_calc_vwc}
\title{Converting soil moisture from TDT signal to volumetric water content}
\usage{
mc_calc_vwc(
  data,
  moist_sensor = "TMS_TMSmoisture",
  temp_sensor = "TMS_T1",
  output_sensor = "vwc_moisture",
  soiltype = "universal",
  localities = NULL,
  t_ref = 24,
  acor_t = 1.91132689118,
  wcor_t = 0.64108
)
}
\arguments{
\item{data}{in format for calculation}

\item{moist_sensor}{name of soil moisture sensor (default "TMS_TMSmoisture")

Soil moisture sensor must be in TMSmoisture physical. If sensor$calibration@intercept
and sensor$calibration@intercept are set, then parameters are used in calculation.
Value sensor$metadata@calibrated is derived from calibration parameters.}

\item{temp_sensor}{name of soil temperature sensor (default "TMS_T1")

Temperature sensor must be in T physical.}

\item{output_sensor}{name of new snow sensor (default "vwc_moisture")}

\item{soiltype}{value from mc_data_vwc_parameters in column soiltype (default "universal")

Parameters a, b and c are used in calculation.}

\item{localities}{list of locality_ids for calculation; if NULL then all (default NULL)}

\item{t_ref}{(default 24)}

\item{acor_t}{(default 1.91132689118)}

\item{wcor_t}{(default 0.64108)}
}
\value{
input data with added VWC moisture sensor
}
\description{
Function converting soil moisture from TDT signal to volumetric water content.
}
\details{

}
