% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{mc_calib_moisture}
\alias{mc_calib_moisture}
\title{Calculation slope and intercept parameters of TMSmoisture sensor}
\usage{
mc_calib_moisture(
  raw_air,
  raw_water,
  t_air = 24,
  t_water = 24,
  ref_air = 114.534,
  ref_water = 3634.723,
  ref_t = .calib_MOIST_REF_T,
  acor_t = .calib_MOIST_ACOR_T,
  wcor_t = .calib_MOIST_WCOR_T
)
}
\arguments{
\item{raw_air}{TDT signal in air}

\item{raw_water}{TDT signal in water}

\item{t_air}{temperature of air (default 24)}

\item{t_water}{temperature of air (default 24)}

\item{ref_air}{(default 114.534)}

\item{ref_water}{(default 3634.723)}

\item{ref_t}{(default 24)}

\item{acor_t}{(default 1.91132689118083)}

\item{wcor_t}{(default 0.64108)}
}
\value{
list with slope and intercept parameters
}
\description{
Function calculate slope and intercept parameters from measuring moisture in water and air.
}
\details{

}
