% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\docType{data}
\name{mc_const_CALIB_MOIST_WCOR_T}
\alias{mc_const_CALIB_MOIST_WCOR_T}
\title{Default temperature drift for TMS moisture in the water}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
mc_const_CALIB_MOIST_WCOR_T
}
\description{
0.64108 = default temperature drift correction parameter in the water -
TMS moisture sensor. This constant is used in the function \link{mc_calc_vwc}.
}
\keyword{datasets}
