% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_Dendro_raw}
\alias{mc_const_SENSOR_Dendro_raw}
\title{Default sensor for TOMST Dendrometer radius difference}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_Dendro_raw
}
\description{
This constant is used in the function \link{mc_calc_tomst_dendro}
as default sensor for converting the change in stem size from raw
TOMST units to micrometers.
mc_const_SENSOR_Dendro_raw = "Dendro_raw"
}
\keyword{datasets}
