% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_TMS_T1}
\alias{mc_const_SENSOR_TMS_T1}
\title{Default sensor for TOMST TMS soil temperature}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_TMS_T1
}
\description{
This constant is used in the function \link{mc_calc_vwc} to account for soil temperature effect
while converting the raw TMS soil moisture (scaled TDT signal) to volumetric water content (VWC).
mc_const_SENSOR_TMS_T1 = "TMS_T1"
}
\keyword{datasets}
