% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{data}
\name{mc_const_SENSOR_TMS_T2}
\alias{mc_const_SENSOR_TMS_T2}
\title{Default sensor for TOMST TMS temperature of soil surface}
\format{
An object of class \code{character} of length 1.
}
\usage{
mc_const_SENSOR_TMS_T2
}
\description{
Default sensor for TOMST TMS temperature of soil surface
}
\keyword{datasets}
