% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_data_example_agg}
\alias{mc_data_example_agg}
\title{Example data in Agg-format.}
\format{
An object of class \code{myClimList} (inherits from \code{list}) of length 2.
}
\usage{
mc_data_example_agg
}
\description{
Cleaned data in Agg-format. Three example localities situated in
Saxon Switzerland National Park.
myClim object has metadata and covers time period from 2020-10 to 2021-02.

Data includes time-series from 4 loggers:
\itemize{
\item Tomst TMS4 with 4 sensors ("TMS_T1", "TMS_T2", "TMS_T3", "TMS_TMSmoisture")
\item Tomst Thermologger with 1 sensor ("TS_T)
\item Tomst Point Dendrometer with 2 sensors ("DEND_T", "DEND_TOMSTdendro")
\item HOBO U23 with 2 sensors ("HOBO_T_C", "HOBO_RH")
}
}
\keyword{datasets}
