% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_data_formats}
\alias{mc_data_formats}
\title{Formats of source data files}
\format{
An object of class \code{environment} of length 3.
}
\usage{
mc_data_formats
}
\description{
R object of class environment with the definitions how to
parse specific microclimatic logger files. In case you would like to add
new, unsupported logger, this is the place where the reading key is stored.
}
\seealso{
\link{mc_DataFormat},\linkS4class{mc_TOMSTDataFormat}, \linkS4class{mc_TOMSTJoinDataFormat}
}
\keyword{datasets}
