% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_data_heights}
\alias{mc_data_heights}
\title{§Default heights of sensors}
\format{
An object of class \code{data.frame} with 13 rows and 4 columns.
}
\usage{
mc_data_heights
}
\description{
This table is used for set height in metadata of sensor during reading source data.
}
\details{
data.frame with columns:
\itemize{
\item logger_type
\item sensor_name
\item height - character representation of height
\item suffix - suffix for sensor_name. If suffix is NA, then sensor_name is original.
}
}
\seealso{
\code{\link[=mc_read_files]{mc_read_files()}}, \code{\link[=mc_read_data]{mc_read_data()}}§
}
\keyword{datasets}
