% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_data_physical}
\alias{mc_data_physical}
\title{Physical quantities definition
R object of class environment with the definitions of physical elements
for recording the microclimate e.g. temperature, speed, depth, volumetric water content...
see \linkS4class{mc_Physical}. Similarly as in case of logger format definitions \linkS4class{mc_DataFormat} it is easy
to add new, unsupported physical here.}
\format{
An object of class \code{environment} of length 8.
}
\usage{
mc_data_physical
}
\description{
Physical quantities definition
R object of class environment with the definitions of physical elements
for recording the microclimate e.g. temperature, speed, depth, volumetric water content...
see \linkS4class{mc_Physical}. Similarly as in case of logger format definitions \linkS4class{mc_DataFormat} it is easy
to add new, unsupported physical here.
}
\seealso{
\linkS4class{mc_Physical}
}
\keyword{datasets}
