% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_data_physical}
\alias{mc_data_physical}
\title{Physical quantities definition}
\format{
An object of class \code{list} of length 2.
}
\usage{
mc_data_physical
}
\description{
Physical quantities definition
}
\keyword{datasets}
