% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_data_physical}
\alias{mc_data_physical}
\title{Physical quantities definition}
\format{
An object of class \code{environment} of length 11.
}
\usage{
mc_data_physical
}
\description{
R object of class environment with the definitions of physical elements
for recording the microclimate e.g. temperature, speed, depth, volumetric water content...
see \linkS4class{mc_Physical}. Similarly as in case of logger format definitions \linkS4class{mc_DataFormat} it is easy
to add new, physical here.
}
\seealso{
\linkS4class{mc_Physical}

Currently supported physical elements:
\itemize{
\item l_cm - length in cm
\item l_mm - length in mm
\item l_um - length in um
\item VWC - volumetric moisture in m3/m3
\item RH - relative humidity in \%
\item T_C - temperature in °C
\item t_h - time in hours
\item moisture_raw - raw TMS moisture sensor values
\item radius_raw - radius difference in raw units
\item v - speed in m/s
}
}
\keyword{datasets}
