% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mc_data_sensors}
\alias{mc_data_sensors}
\title{Sensors definition}
\format{
An object of class \code{environment} of length 17.
}
\usage{
mc_data_sensors
}
\description{
Sensors definition
}
\keyword{datasets}
