% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco.R
\name{mc_eco_snow_agg}
\alias{mc_eco_snow_agg}
\title{Snow detection summary}
\usage{
mc_eco_snow_agg(data, sensor, localities = c(), dr = 2, tmax = 0.5, period = 3)
}
\arguments{
\item{data}{all data in standard format}

\item{sensor}{name of temperature sensor}

\item{localities}{names of localities; if empty then all}

\item{dr}{delta range}

\item{tmax}{maximal temperature}

\item{period}{count days for continuous cover of snow (default 3)}
}
\value{
data.frame with columns serial_number, snow_days, first_day, last_day, first_day_period, last_day_period
}
\description{
Function return summary info about snow detection
}
\examples{
mc_eco_snow_agg(example_tomst_data1, "TMS_T3")
}
