% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{mc_feed_directory}
\alias{mc_feed_directory}
\title{Reading files from directory}
\usage{
mc_feed_directory(directory, logger_type, recursive = TRUE)
}
\arguments{
\item{directory}{character}

\item{logger_type}{character - type of logger (TMS)}

\item{recursive}{logical - recursive search in subdirectories}
}
\value{
data in standard format
}
\description{
This function read data files from directory of one logger type. Locality is set None.
}
\examples{
example_tms_data <- microclim::mc_feed_directory("examples/data/TMS/", "TMS")
}
