% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{mc_feed_files}
\alias{mc_feed_files}
\title{Reading files}
\usage{
mc_feed_files(files, dataformat_name)
}
\arguments{
\item{files}{vector of character - files with data}

\item{dataformat_name}{character - data format of logger (TOMST)}
}
\value{
data in standard format
}
\description{
This function read data files of one logger type. Locality is set None.
}
\examples{
example_tomst_data <- microclim::mc_feed_files(c("examples/data/TOMST/data_91184101_0.csv", "examples/data/TOMST/data_94184102_0.csv"), "TOMST")
}
