% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{mc_filter}
\alias{mc_filter}
\title{Filter data from myClim object}
\usage{
mc_filter(
  data,
  localities = NULL,
  sensors = NULL,
  reverse = FALSE,
  stop_if_empty = TRUE
)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{localities}{locality_ids for filtering data; if NULL then do nothing}

\item{sensors}{sensor_ids for filtering data; if NULL then do nothing see \code{names(mc_data_sensors)}}

\item{reverse}{if TRUE then input localities and/or sensors are excluded (default FALSE)}

\item{stop_if_empty}{if TRUE then error for empty output (default TRUE)}
}
\value{
filtered myClim object
}
\description{
This function filter data by localities and sensors.
}
\details{
In default settings it returns the object containing input sensors / localities.
It is possible to filter selected sensors from all localities, selected
localities with all sensors or combine both.

When \code{reverse = TRUE} then
resulting myClim object contains all localities and sensors as input except
those you listed in parameters \verb{localities, sensors}
}
\examples{
filtered_data <- mc_filter(mc_data_example_raw, localities=c("A6W79", "A2E32"),
                           sensors=c("TMS_T1", "TMS_T2"))
mc_info(filtered_data)
}
