% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mc_info}
\alias{mc_info}
\title{Get sensors info table}
\usage{
mc_info(data)
}
\arguments{
\item{data}{myClim object in Prep-format or Calc-format (see \link{myClim-package})}
}
\value{
data.frame with columns:
\itemize{
\item locality_id
\item serial_number
\item sensor_id
\item sensor_name
\item start_date - the oldest record on the sensor
\item end_date - the newest record on the sensor
\item step - time step of records series (minutes)
\item step_text - time step of records series (text)
\item min_value - minimal recorded values
\item max_value - maximal recorded value
\item count_values - number of non NA records
\item count_na - number of NA records
}
}
\description{
This function return data.frame with info about sensors
}
