% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mc_info_clean}
\alias{mc_info_clean}
\title{Call cleaning log}
\usage{
mc_info_clean(data)
}
\arguments{
\item{data}{myClim object in Raw-format. see \link{myClim-package}}
}
\value{
data.frame with columns:
\itemize{
\item locality_id - when provided by user then locality ID, when not provided identical with serial number
\item serial_number - serial number of logger when provided or automatically detected from file name or header
\item start_date - date of the first record on the logger
\item end_date  - date of the last record on the logger
\item step_seconds - detected time step in seconds of the logger measurements.
\item count_duplicities - number of duplicated records (identical time)
\item count_missing - number of missing records (logger outage in time when it should record)
\item count_disordered - number of records incorrectly ordered in time (newer followed by older)
\item rounded - T/F indication whether myClim automatically rounded time series minutes to the closes half (HH:00, HH:30) e.g. 13:07 -> 13:00
}
}
\description{
This function return data.frame with information from cleaning the loggers time series see \code{\link[=mc_prep_clean]{mc_prep_clean()}}
}
\seealso{
\code{\link[=mc_prep_clean]{mc_prep_clean()}}
}
