% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{mc_info_count}
\alias{mc_info_count}
\title{Count data}
\usage{
mc_info_count(data)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}
}
\value{
data.frame with count of localities, loggers and sensors
}
\description{
This function return data.frame with the number of localities, loggers and sensors of input myClim object.
}
\examples{
count_table <- mc_info_count(mc_data_example_raw)
}
