% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{mc_join}
\alias{mc_join}
\title{Joining time-series from repeated downloads}
\usage{
mc_join(data, comp_sensors = NULL)
}
\arguments{
\item{data}{myClim object in Prep-format. See \link{myClim-package}}

\item{comp_sensors}{senors for compare and select source logger; If NULL then first is used. (default NULL)}
}
\value{
myClim object with joined loggers.
}
\description{
Function join time-series from repeated download on the same locality. It joins fragmentary
time-series from the downloads of certain locality and identical loggers type \link{mc_LoggerMetadata} and time step.
While joining, each sensor from older downloaded logger must have pair sensor on
the newer downloaded logger with same height. If time-series overlapps and values are different,
then user interactively select which source logger to use.
Plot with conflict interval is shown during interactive joining. User can select older time-series,
newer time-series or set manually the datetime split of newer and older.
}
\details{
Joining is semi-automatic pairwise process. myClim is sequentially joining the oldest logger
file with the newer. In case the time series fits well (the newer starts at the point where the older ends),
or in case the time series overlap and values on the overlap are identical, then myClim join automatically.
Otherwise user interactively selects which time series to use.

Loggers with multiple sensors are joined base on one or more selected sensors. See parameter \code{comp_sensors}.

Name of result sensor is used from logger with older data. If \code{serial_number} is not equal while joining loggers, then
the result \code{serial_number} is \code{NA}. Clean info is changed to \code{NA} except step. If  not calibrated sensor is beeing
joined with calibrated one, then calibration information must be empty in not calibrated sensor.
}
