% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{mc_join}
\alias{mc_join}
\title{Joining sensors from different loggers}
\usage{
mc_join(data)
}
\arguments{
\item{data}{myClim object in Prep-format. See \link{myClim-package}}
}
\value{
myClim object with joined sensors.
}
\description{
Function join sensors from different loggers. Loggers with same type \link{mc_LoggerMetadata} and step are joined.
Every sensore from first logger must have pair sensor from second one with same height.
}
\details{
Name of result sensor is used from logger with older data. If serial_number is not equal in joining loggers, then
result serial_number is NA. Clean info is changed to NA except step. If uncalibrated sensor is joining with calibrated one,
then calibration inforamtion must be empty for uncalibrated sensor.
}
