% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{mc_join}
\alias{mc_join}
\title{Joining time-series from repeated downloads}
\usage{
mc_join(data, comp_sensors = NULL)
}
\arguments{
\item{data}{myClim object in Raw-format. see \link{myClim-package}}

\item{comp_sensors}{senors for compare and select source logger; If NULL then first is used. (default NULL)}
}
\value{
myClim object with joined loggers.
}
\description{
Function join time-series from repeated download on the same locality.
If time-series overlaps and values are different,
then user interactively select which source logger to use.
Plot with conflict interval is shown during interactive joining.
}
\details{
Joining is semi-automatic pairwise process.
In case the end of older and start of the newer fits
myClim is sequentially joining the oldest logger file with the newer.
In case the time series overlap and values on the overlap are not identical,
user interactively selects which time series to us on the
plot with conflict interval.

Loggers with multiple sensors are joined base on one or more selected sensors.
See parameter \code{comp_sensors}.

Name of result sensor is used from logger with older data. If \code{serial_number}
is not equal while joining loggers, then the result \code{serial_number} is \code{NA}.
Clean info is changed to \code{NA} except step. If  not calibrated sensor is being
joined with calibrated one, then calibration information must be empty in not calibrated sensor.
}
