% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{mc_load}
\alias{mc_load}
\title{Load myClim object}
\usage{
mc_load(file)
}
\arguments{
\item{file}{path to input RDS file}
}
\value{
loaded myClim object
}
\description{
§This function load the myClim data object from an RDS file. The standard R functions load and readRDS are unsafe
to load the MyClim object. The mc_save a mc_load functions ensure that the myClim object is correctly loaded
in the newer version.§
}
