% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_LocalityMetadata-class}
\alias{mc_LocalityMetadata-class}
\alias{mc_LocalityMetadata}
\title{Class for locality metadata}
\description{
Class for locality metadata
}
\details{
When reading without metadata, then locality is named after file
where the data come from, or after the sensor id where the data come form.
}
\section{Slots}{

\describe{
\item{\code{locality_id}}{name of locality}

\item{\code{elevation}}{of locality}

\item{\code{lat_wgs84}}{latitude of locality in WGS-84}

\item{\code{lon_wgs84}}{longitude of locality in WGS-84}

\item{\code{tz_offset}}{offset from UTC in minutes}

\item{\code{tz_type}}{type of time zone}

\item{\code{user_data}}{list for user data}
}}

\seealso{
\link{myClim-package}, \link{mc_LoggerMetadata}, \link{mc_SensorMetadata}
}
