% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_LoggerCleanInfo-class}
\alias{mc_LoggerCleanInfo-class}
\alias{mc_LoggerCleanInfo}
\title{Class for logger clean info}
\description{
Class for logger clean info
}
\section{Slots}{

\describe{
\item{\code{step}}{of series in minutes}
}}

