% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_LoggerCleanInfo-class}
\alias{mc_LoggerCleanInfo-class}
\alias{mc_LoggerCleanInfo}
\title{Class for logger clean info}
\description{
Class for logger clean info
}
\section{Slots}{

\describe{
\item{\code{step}}{Time step of microclimatic data series in seconds}

\item{\code{count_duplicities}}{count of duplicated records - values with same date}

\item{\code{count_missing}}{count of missing records; Period between the records should be the same length. If not, than missing.}

\item{\code{count_disordered}}{count of records incorrectly ordered in time. In table, newer record is followed by the older.}

\item{\code{rounded}}{T/F indication whether myClim automatically rounded time series to the closes half (HH:00, HH:30) e.g. 13:07 -> 13:00}
}}

