% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_LoggerMetadata-class}
\alias{mc_LoggerMetadata-class}
\alias{mc_LoggerMetadata}
\title{Class for logger metadata}
\description{
Class for logger metadata
}
\section{Slots}{

\describe{
\item{\code{type}}{of logger (TMS, ThermoDatalogger)}
}}

