% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_LoggerMetadata-class}
\alias{mc_LoggerMetadata-class}
\alias{mc_LoggerMetadata}
\title{Class for logger metadata}
\description{
Class for logger metadata
}
\section{Slots}{

\describe{
\item{\code{type}}{of logger (TMS, ThermoDatalogger, Dendrometer, HOBO)}

\item{\code{serial_number}}{serial number of the logger}

\item{\code{step}}{time step of microclimatic time-seris in seconds.
When provided by user, is used in \code{\link[=mc_prep_clean]{mc_prep_clean()}} function instead of
automatic step detection}
}}

