% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_MainMetadata-class}
\alias{mc_MainMetadata-class}
\alias{mc_MainMetadata}
\title{Class for main metadata in data format for calculation}
\description{
Class for main metadata in data format for calculation
}
\section{Slots}{

\describe{
\item{\code{step}}{of data}
}}

