% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_MainMetadata-class}
\alias{mc_MainMetadata-class}
\alias{mc_MainMetadata}
\title{Class for myClim object metadata}
\description{
Class for myClim object metadata
}
\section{Slots}{

\describe{
\item{\code{version}}{§of the myClim packege in which was created§}

\item{\code{format_type}}{§type of format (raw, agg)§}
}}

\seealso{
\link{myClim-package}
}
