% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_MainMetadata-class}
\alias{mc_MainMetadata-class}
\alias{mc_MainMetadata}
\title{Class for myClim object metadata in Calc-format}
\description{
Class for myClim object metadata in Calc-format
}
\section{Slots}{

\describe{
\item{\code{step}}{time step of data in minutes}

\item{\code{step_text}}{time step of data in text for irregular periods (e.g. month)}
}}

\seealso{
Calc-format section of \link{myClim-package}
}
