% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_MainMetadata-class}
\alias{mc_MainMetadata-class}
\alias{mc_MainMetadata}
\title{Class for myClim object metadata}
\description{
Class for myClim object metadata
}
\section{Slots}{

\describe{
\item{\code{version}}{the version of the myClim package in which the object was created}

\item{\code{format_type}}{type of format (Raw-format, Agg-format)}
}}

\seealso{
\link{myClim-package}
}
