% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_Physical-class}
\alias{mc_Physical-class}
\alias{mc_Physical}
\title{Class for physical}
\description{
Class defining the element of the records (temperature, volumetric water content, height...)
}
\details{
See e.g. definition of temperature. Similarly as the definition of new loggers, new
physicals can be added like modules.
\preformatted{
Slot "name": "T_C"
Slot "description": "Temperature °C"
Slot "units": "°C"
Slot "viridis_color_map": "C" 
Slot "scale_coeff": 0.03333333
}
}
\section{Slots}{

\describe{
\item{\code{name}}{of physical}

\item{\code{description}}{character info}

\item{\code{units}}{measurument (°C, \%, m3/m3, raw, mm, ...)}

\item{\code{viridis_color_map}}{viridis color map option}

\item{\code{scale_coeff}}{coefficient for plot; value * scale_coef is in range 0-1}
}}

\seealso{
\link{mc_data_physical}
}
