% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mc_plot_image}
\alias{mc_plot_image}
\title{Plot data - image}
\usage{
mc_plot_image(
  data,
  filename,
  title,
  localities = NULL,
  sensors = NULL,
  height = 1900,
  left_margin = 12
)
}
\arguments{
\item{data}{all data in standard format}

\item{filename}{output filename}

\item{title}{of plot}

\item{localities}{names of localities; if empty then all}

\item{sensors}{names of sensors; if empty then all}

\item{height}{of image; default = 1900}

\item{left_margin}{width of space for sensor_labels; default = 12}
}
\description{
Function plot data to file with image function
}
\examples{
mc_plot_image(data, "T1_image.png", "T1 sensor", sensors="TMS_T1")
}
