% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mc_plot_image}
\alias{mc_plot_image}
\title{Plot data - image}
\usage{
mc_plot_image(
  data,
  filename,
  title = "",
  localities = NULL,
  sensors = NULL,
  height = 1900,
  left_margin = 12
)
}
\arguments{
\item{data}{myClim object in Prep-format or Calc-formt see \link{myClim-package}}

\item{filename}{output file name (file path)}

\item{title}{of plot; default is empty}

\item{localities}{filter the localities; if empty then all}

\item{sensors}{filter the sensors; if empty then all see \code{names(mc_data_sensors)}}

\item{height}{of image; default = 1900}

\item{left_margin}{width of space for sensor_labels; default = 12}
}
\description{
Function plot myClim data into file with image function.
}
\details{
Be careful with bigger data. Can take some time.
}
\examples{
\dontrun{mc_plot_image(data, "T1_image.png", "T1 sensor", sensors="TMS_T1")}
}
