% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mc_plot_image}
\alias{mc_plot_image}
\title{Plot data}
\usage{
mc_plot_image(data_table, filename, title)
}
\arguments{
\item{data_table}{data.frame in wideformat generated by mc_reshape_wideformat_interval}

\item{filename}{output filename}
}
\description{
Function plot data to file
}
\examples{
data_table <- mc_reshape_wideformat_interval(example_tms_data1, sensors="T1")
mc_plot_image(data_table, "T1_image.png", "T1 sensor")
}
