% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mc_plot_line}
\alias{mc_plot_line}
\title{Plot data - ggplot2 geom_line}
\usage{
mc_plot_line(
  data,
  filename,
  sensors = NULL,
  scale_coeff = NULL,
  png_width = 1900,
  png_height = 1900,
  start_crop = NULL,
  end_crop = NULL
)
}
\arguments{
\item{data}{myClim object in Prep-format or Calc-formt see \link{myClim-package}}

\item{filename}{output - supported formats are pdf and png}

\item{sensors}{select the names of sensors to be plotted (max 2) see \code{names(mc_data_sensors)}}

\item{scale_coeff}{scale coefficient for secondary axis (default NULL)}

\item{png_width}{width for png output (default 1900)}

\item{png_height}{height for png output (default 1900)}

\item{start_crop}{POSIXct datetime for crop data (default NULL)}

\item{end_crop}{POSIXct datetime for crop data (default NULL)}
}
\description{
Function plot data to file with ggplot2 geom_line
}
\details{
Maximal number of physical units (elements) of sensors to be plotted  is two. Main and secondary y axis.
Values from secondary axis are scaled with calculation values * scale_coeff. If coefficient is NULL
than function try detects scale coefficient from physical unit of sensors see \linkS4class{mc_Physical}.
Scaling is useful when plotting together e.g. temperature and moisture.
}
