% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mc_plot_loggers}
\alias{mc_plot_loggers}
\title{Plot data from loggers}
\usage{
mc_plot_loggers(
  data,
  directory,
  localities = c(),
  sensors = c(),
  crop = c(NA, NA)
)
}
\arguments{
\item{data}{all data in standard format}

\item{directory}{output directory}

\item{localities}{names of localities; if empty then all}

\item{sensors}{names of sensors; if empty then all}

\item{crop}{datetime range for plot, not cropping if NA (default c(NA, NA))}
}
\description{
Function plot loggers to directory
}
\examples{
mc_plot_loggers(example_tomst_data1, "Figures")
}
