% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mc_plot_loggers}
\alias{mc_plot_loggers}
\title{Plot data from loggers}
\usage{
mc_plot_loggers(
  data,
  directory,
  localities = NULL,
  sensors = NULL,
  crop = c(NA, NA)
)
}
\arguments{
\item{data}{myClim object in Raw-format. see \link{myClim-package}}

\item{directory}{output directory}

\item{localities}{names of localities; if NULL then all (default NULL)}

\item{sensors}{names of sensors; if NULL then all (default NULL) see \code{names(mc_data_sensors)}}

\item{crop}{datetime range for plot, not cropping if NA (default c(NA, NA))}
}
\description{
Function save separate file for the loggers to the directory. Only Raw-format supported.
For Agg-format use \code{\link[=mc_plot_line]{mc_plot_line()}}
}
\examples{
\dontrun{mc_plot_loggers(example_tomst_data1, "Figures")}
}
