% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mc_plot_raster}
\alias{mc_plot_raster}
\title{Plot data - ggplot2 geom_raster}
\usage{
mc_plot_raster(
  data,
  filename = NULL,
  sensors = NULL,
  by_hour = TRUE,
  png_width = 1900,
  png_height = 1900,
  viridis_color_map = NULL,
  start_crop = NULL,
  end_crop = NULL
)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{filename}{output with the extension - supported formats are .pdf and .png (default NULL)

If NULL then the plot is return, but not saved to file.}

\item{sensors}{names of sensor; should have same unit see \code{names(mc_data_sensors)}}

\item{by_hour}{if TRUE, then y axis is plotted as an hour, else original time step (default TRUE)}

\item{png_width}{width for png output (default 1900)}

\item{png_height}{height for png output (default 1900)}

\item{viridis_color_map}{viridis color map option; if NULL, then used value from mc_data_physical
\itemize{
\item "A" - magma
\item "B" - inferno
\item "C" - plasma
\item "D" - viridis
\item "E" - cividis
\item "F" - rocket
\item "G" - mako
\item "H" - turbo
}}

\item{start_crop}{POSIXct datetime for crop data (default NULL)}

\item{end_crop}{POSIXct datetime for crop data (default NULL)}
}
\value{
list of ggplot2 objects
}
\description{
Function plot data to file with ggplot2 geom_raster
}
\details{
PDF format is recommended as it can distribute the plots on the pages, which is especially useful
for bigger data. In case of plotting multiple sensors, it is plotted by sensor. All localities from sensor1
followed by all localities of sensor2 etc. When plotting only few localities, but multiple sensors,
each sensor has own page. E.g. when plotting data from one locality, and 3 senosrs resulting PDF has 3 pages.
In case of plotting PNG all loclaities and sensors are plotted in one image. Be careful with bigger data in PNG.
Play with \code{png_height} and \code{png_width}. When too small, image does not fit ad is plotted broken.
}
