% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mc_plot_raster}
\alias{mc_plot_raster}
\title{Plot data - ggplot2 geom_raster}
\usage{
mc_plot_raster(
  data,
  filename,
  sensors = NULL,
  by_hour = TRUE,
  png_width = 1900,
  png_height = 1900,
  viridis_color_map = NULL,
  start_crop = NULL,
  end_crop = NULL
)
}
\arguments{
\item{data}{in format for preparing or calculation}

\item{filename}{output - supported formats are pdf and png}

\item{sensors}{names of sensor; should have same unit}

\item{by_hour}{if TRUE, then y axis is hour, alse time (default TRUE)}

\item{png_width}{width for png output (default 1900)}

\item{png_height}{height for png output (default 1900)}

\item{viridis_color_map}{viridis color map option; if NULL, then used value from mc_data_physical
\itemize{
\item "A" - magma
\item "B" - inferno
\item "C" - plasma
\item "D" - viridis
\item "E" - cividis
\item "F" - rocket
\item "G" - mako
\item "H" - turbo
}}

\item{start_crop}{POSIXct datetime for crop data (default NULL)}

\item{end_crop}{POSIXct datetime for crop data (default NULL)}
}
\description{
Function plot data to file with ggplot2 geom_raster
}
