% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_calib_load}
\alias{mc_prep_calib_load}
\title{load calibration parameters}
\usage{
mc_prep_calib_load(data, calib_table)
}
\arguments{
\item{data}{in format for preparing}

\item{calib_table}{data.frame with columns (serial_number, sensor_id, datetime, slope, intercept)}
}
\value{
data with loaded calibration informations.
}
\description{
This function load calibration parameters from data.frame
}
\details{
It is not possible change calibration parameters in calibrated sensor.
}
