% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_clean}
\alias{mc_prep_clean}
\title{Cleaning datetime series}
\usage{
mc_prep_clean(data, silent = FALSE)
}
\arguments{
\item{data}{in format for preparing}

\item{silent}{if true, then informations aren't printed (default FALSE)}
}
\value{
cleaned data in standard format
}
\description{
This function change datetime and values series. Result series has constant
step without duplicits and missed values are filled in as NA.
}
\examples{
cleaned_example_tomst_data1 <- mc_prep_clean(example_tomst_data1)
}
