% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_crop}
\alias{mc_prep_crop}
\title{Crop datetime}
\usage{
mc_prep_crop(data, start = NULL, end = NULL, end_included = TRUE)
}
\arguments{
\item{data}{in format for preparing or calculation}

\item{start}{POSIXct datetime in UTC; is optional; start datetime is included}

\item{end}{POSIXct datetime in UTC; is optional}

\item{end_included}{if TRUE then  end datetime is included (default TRUE)}
}
\value{
cropped data in standard format
}
\description{
This function crop data by datetime
}
\examples{
cleaned_example_tomst_data1 <- mc_prep_crop(example_tomst_data1, end=as.POSIXct("2020-02-01", tz="UTC"))
}
