% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_crop}
\alias{mc_prep_crop}
\title{Crop datetime}
\usage{
mc_prep_crop(data, start = NULL, end = NULL, end_included = TRUE)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{start}{POSIXct datetime in UTC; is optional; start datetime is included}

\item{end}{POSIXct datetime in UTC; is optional}

\item{end_included}{if TRUE then end datetime is included (default TRUE)}
}
\value{
cropped data in the same myClim format as input.
}
\description{
This function crop data by datetime
}
\details{
Function is able to crop data from start to end but works also
with only start and only end. When only start provided, then crop only start and
do not touch the end and vice versa.
}
\examples{
cropped_data <- mc_prep_crop(mc_data_example_clean, end=as.POSIXct("2020-02-01", tz="UTC"))
}
