% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_merge}
\alias{mc_prep_merge}
\title{Merge myClim objects}
\usage{
mc_prep_merge(data_items)
}
\arguments{
\item{data_items}{list of myClim objects in Prep-format or Calc-format see \link{myClim-package}; Format of merged objects must be same.}
}
\value{
merged myClim object in the same format as input objects
}
\description{
This function is designed to merge two myClim objects into one.
}
\details{
This function works only when the input myClim objects have the same format (Prep-format, Calc-format) and the same time step.
This function do not join the time component of the data but is primary designed to merge different localities.
If data1 and data2 contains locality with same names (locality_id), than locality_id from data2 is renamed, data on identical localities are not merged.
}
\examples{
merged_tomst_data <- mc_prep_merge(list(example_tomst_data1, example_tomst_data2))
}
