% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_merge}
\alias{mc_prep_merge}
\title{Merge data}
\usage{
mc_prep_merge(data_items)
}
\arguments{
\item{data_items}{list of data in format for preparing or calculation; Format of data must be same.}
}
\value{
merged data
}
\description{
This function merge two instances of data to one
}
\details{
If data1 and data2 contains locality with same locality_id, than locality_id from data2 is renamed.
}
\examples{
merged_tomst_data <- mc_prep_merge(list(example_tomst_data1, example_tomst_data2))
}
