% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{mc_prep_rename_sensor}
\alias{mc_prep_rename_sensor}
\title{Rename sensor}
\usage{
mc_prep_rename_sensor(
  data,
  sensor_names,
  localities = NULL,
  serial_numbers = NULL
)
}
\arguments{
\item{data}{in format for preparing or calculation}

\item{sensor_names}{list with new names of sensors; names of items are old ones}

\item{localities}{vector of locality_ids; if NULL than all (default NULL)}

\item{serial_numbers}{vector of serial_numbers; if NULL than all (default NULL); parameter is usefull only for
preparing format of data}
}
\value{
data with changed sensor names
}
\description{
This function rename sensors. It is usefull for flatting data format.
}
