% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{mc_read_data}
\alias{mc_read_data}
\title{Data files reading}
\usage{
mc_read_data(files_table, localities_table = NULL)
}
\arguments{
\item{files_table}{csv file path or data.frame which describe data files

Csv file required header row. Column separator is ",".

Columns:
\itemize{
\item path - path to file
\item locality_id
\item data_format
\item serial_number - can be NA, than try detect
}}

\item{localities_table}{csv file path or data.frame which describe localities

Csv file required header row. Column separator is ",".

Columns:
\itemize{
\item locality_id
\item altitude
\item lon_wgs84
\item lat_wgs84
\item tz_offset
}}
}
\value{
data in standard format
}
\description{
This function read raw data from loggers by data.frame or csv file with files description.
}
\examples{
\dontrun{
tomst_data <- mc_read_data("examples/data/TOMST/files_table.csv", "examples/data/TOMST/localities_table.csv")
}
}
