% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{mc_read_files}
\alias{mc_read_files}
\title{Reading files or directories}
\usage{
mc_read_files(
  paths,
  dataformat_name,
  recursive = TRUE,
  date_format = NA_character_,
  tz_offset = NA_integer_
)
}
\arguments{
\item{paths}{vector of paths to files or directories}

\item{dataformat_name}{character - data format of logger one of (TOMST, TOMST_join) see \code{names(mc_data_formats)}}

\item{recursive}{logical - recursive search in sub-directories (default TRUE)}

\item{date_format}{\itemize{
\item ||format of date used in strptime function e.g. "\%d.\%m.\%y \%H:\%M:\%S" (default NA) necessary to provide for HOBO.
}

This parameter is required only for variable data formats as HOBO. Date format in TOMST data is stable.
see \link{mc_data_formats}||}

\item{tz_offset}{\itemize{
\item ||timezone offset in minutes; It is required fill in only for non-UTC source data. (default NA)||
}}
}
\value{
myClim object in Prep-format see \link{myClim-package}
}
\description{
This function read one or more csv files or directories of identical,
pre-defined logger type (format) see \link{mc_DataFormat} and \link{mc_data_formats}.
This function does not allow you to provide additional locality or sensor metadata.
Metadata can be load through \code{\link[=mc_read_data]{mc_read_data()}} or
can be provided later with function \code{\link[=mc_prep_meta]{mc_prep_meta()}}
}
\details{
If file is not in expected format, then file is skipped and warning printed in console.
CSV files (loggers raw data) are in resulting myClim object placed to separate localities with empty metadata.
Localities are named after serial_number of logger.
}
\examples{
\dontrun{
tomst_data <- mc_read_files(c("examples/data/TOMST/data_91184101_0.csv", "examples/data/TOMST/data_94184102_0.csv"), "TOMST")
}
}
\seealso{
\link{mc_DataFormat}
}
