% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{mc_read_files}
\alias{mc_read_files}
\title{Reading files or directories}
\usage{
mc_read_files(
  paths,
  dataformat_name,
  recursive = TRUE,
  date_format = NA_character_,
  tz_offset = NA_integer_,
  step = NA_integer_
)
}
\arguments{
\item{paths}{vector of paths to files or directories}

\item{dataformat_name}{data format of logger one of \code{names(mc_data_formats)} see \code{names(mc_data_formats)}}

\item{recursive}{recursive search in sub-directories (default TRUE)}

\item{date_format}{format of date in your hobo files e.g. "\%d.\%m.\%y \%H:\%M:\%S" (default NA). Required for HOBO files.
For TMS files ignored, there is fix date format. see \link{mc_data_formats}}

\item{tz_offset}{timezone offset in minutes; It is required fill only for non-UTC data (custom settings in HOBO). Not used in TMS (default NA)}

\item{step}{Time step of microclimatic time-series in minutes. When provided, then is used in \link{mc_prep_clean} instead of automatic stepd detection.
If not provided (NA), is automatically detected in \link{mc_prep_clean}. (default NA)}
}
\value{
myClim object in Prep-format see \link{myClim-package}
}
\description{
This function read one or more csv files or directories of identical,
pre-defined logger type (format) see \link{mc_DataFormat} and \link{mc_data_formats}.
This function does not allow you to provide additional locality or sensor metadata.
Metadata can be load through \code{\link[=mc_read_data]{mc_read_data()}} or
can be provided later with function \code{\link[=mc_prep_meta]{mc_prep_meta()}}
}
\details{
If file is not in expected format, then file is skipped and warning printed in console.
CSV files (loggers raw data) are in resulting myClim object placed to separate localities with empty metadata.
Localities are named after serial_number of logger.
}
\examples{
\dontrun{
tomst_data <- mc_read_files(c("examples/data/TOMST/data_91184101_0.csv", "examples/data/TOMST/data_94184102_0.csv"), "TOMST")
}
}
\seealso{
\link{mc_DataFormat}
}
