% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{mc_read_files}
\alias{mc_read_files}
\title{Reading files or directories}
\usage{
mc_read_files(
  paths,
  dataformat_name,
  logger_type = NA_character_,
  recursive = TRUE,
  date_format = NA_character_,
  tz_offset = NA_integer_,
  step = NA_integer_,
  clean = TRUE,
  silent = FALSE
)
}
\arguments{
\item{paths}{vector of paths to files or directories}

\item{dataformat_name}{data format of logger one of \code{names(mc_data_formats)}}

\item{logger_type}{type of logger (default NA), can be one of
pre-defined see \code{\link[=mc_read_data]{mc_read_data()}} or any custom string}

\item{recursive}{recursive search in sub-directories (default TRUE)}

\item{date_format}{format of date in your hobo files e.g. "\%d.\%m.\%y \%H:\%M:\%S" (default NA).
Required for HOBO files. For TMS files ignored, there is known, stable date format. see \link{mc_data_formats}}

\item{tz_offset}{timezone offset in minutes; It is required only for non-UTC data
(custom settings in HOBO). Not used in TMS (default NA)}

\item{step}{time step of microclimatic time-series in seconds. When provided, then is used in
\link{mc_prep_clean} instead of automatic step detection.
If not provided (NA), is automatically detected in \link{mc_prep_clean}. (default NA)}

\item{clean}{if TRUE, then \link{mc_prep_clean} is called automatically while reading (default TRUE)}

\item{silent}{if TRUE, then any information is printed in console (default FALSE)}
}
\value{
myClim object in Raw-format see \link{myClim-package}
}
\description{
This function read one or more csv files or directories of identical,
pre-defined logger type (format) see \link{mc_DataFormat} and \link{mc_data_formats}.
This function does not support loading locality or sensor metadata while reading.
Metadata can be loaded through \code{\link[=mc_read_data]{mc_read_data()}} or
can be provided later with function \code{\link[=mc_prep_meta_locality]{mc_prep_meta_locality()}}
}
\details{
If file is not in expected format, then file is skipped and warning printed in console.
CSV files (loggers raw data) are in resulting myClim object placed to separate
localities with empty metadata. Localities are named after serial_number of logger.
Pre-defined logger types are ("Dendrometer","HOBO","ThermoDatalogger","TMS","TMS_L45")
By default data are cleaned with function \code{\link[=mc_prep_clean]{mc_prep_clean()}}. See function description. It detects
holes in time-series, duplicated records or records in wrong order.
}
\examples{
files <- c(system.file("extdata", "data_91184101_0.csv", package = "myClim"),
           system.file("extdata", "data_94184102_0.csv", package = "myClim"))
tomst_data <- mc_read_files(files, "TOMST")
}
\seealso{
\link{mc_DataFormat}, \code{\link[=mc_prep_clean]{mc_prep_clean()}}
}
