% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{mc_read_long}
\alias{mc_read_long}
\title{Reading data from long data.frame}
\usage{
mc_read_long(data_table, sensor_ids)
}
\arguments{
\item{data_table}{long data.frame with Columns:
\itemize{
\item locality_id
\item sensor_name
\item datetime - POSIXct and UTC timezone is required
\item value
}}

\item{sensor_ids}{list with relations between sensor_names and sensor_ids;
Sensor_id is key from \link{mc_data_sensors}
\code{sensor_ids <- list(sensor_name1=sensor_id1, sensor_name2=sensor_id2)}}
}
\value{
myClim object in Prep-format
}
\description{
This is universal function designed to read time series and values
from long data.frame to myClim object.
}
\details{
Similar like \link{mc_read_wide} but is capable to read multiple sensors.
}
\seealso{
\link{mc_read_wide}
}
