% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{mc_reshape_long}
\alias{mc_reshape_long}
\title{Export values to long table}
\usage{
mc_reshape_long(data, localities = NULL, sensors = NULL)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{localities}{names of localities; if NULL then all (default NULL)}

\item{sensors}{names of sensors; if NULL then all (default NULL) see \code{names(mc_data_sensors)}}
}
\value{
data.frame

columns:
\itemize{
\item locality_id
\item serial_number
\item sensor_name
\item height
\item datetime
\item time_to
\item value
}
}
\description{
This function converts myClim object to long R data.frame.
}
\examples{
head(mc_reshape_long(mc_data_example_clean, c("A6W79", "A2E32"), c("TMS_T1", "TMS_T2")), 10)
}
