% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{mc_reshape_wide}
\alias{mc_reshape_wide}
\title{Wideformat of sensor values}
\usage{
mc_reshape_wide(data, localities = NULL, sensors = NULL)
}
\arguments{
\item{data}{myClim object in Prep-format or Calc-format (see \link{myClim-package})}

\item{localities}{names of localities; if NULL then all (default NULL)}

\item{sensors}{names of sensors; if NULL then all (default NULL)}
}
\value{
data.frame with columns:
\itemize{
\item datetime
\item locality1_sensor1
\item ...
\item ...
\item localityn_sensorn
}
}
\description{
This function converts myClim object to data.frame with values of sensor in wide format.
}
\details{
First columnt of the output data.frame is datetime followed by the
columns for every sensor. Name of the column is in format:
\itemize{
\item localityid_serialnumber_sensorname for Prep-format
\item localityid_sensorname for Calc-format
}
}
\examples{
example_tms_wideformat <- mc_reshape_wide(example_tomst_data1, c("A6W79", "A2E32"), c("TMS_T1", "TMS_T2"))
}
