% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{mc_reshape_wide}
\alias{mc_reshape_wide}
\title{Wideformat of sensor values}
\usage{
mc_reshape_wide(data, localities = NULL, sensors = NULL)
}
\arguments{
\item{data}{in format for preparing or calculation}

\item{localities}{names of localities; if NULL then all (default NULL)}

\item{sensors}{names of sensors; if NULL then all (default NULL)}
}
\value{
data.frame with datetime column and columns for every sensor; Name of column is in format
{locality_id}\emph{{serial_number}}{sensor_name} for preparing format and
{locality_id}_{sensor_name} for calculation format.
}
\description{
This function create data.frame with values of sensor in wide format.
}
\examples{
example_tms_wideformat <- mc_reshape_wide(example_tomst_data1, c("A6W79", "A2E32"), c("TMS_T1", "TMS_T2"))
}
