% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{mc_reshape_wide}
\alias{mc_reshape_wide}
\title{Export values to wide table}
\usage{
mc_reshape_wide(
  data,
  localities = NULL,
  sensors = NULL,
  use_utc = TRUE,
  show_logger_name = FALSE
)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{localities}{names of localities; if NULL then all (default NULL)}

\item{sensors}{names of sensors; if NULL then all (default NULL) see \code{names(mc_data_sensors)}}

\item{use_utc}{if FALSE, then the time shift from \code{tz_offset} metadata is used to correct
(shift) the output time-series (default TRUE)

In the Agg-format myClim object \code{use_utc = FALSE} is allowed only for steps shorter than one day. In myClim
the day nd longer time steps are defined by the midnight, but this represent whole day, week, month, year...
shifting daily, weekly, monthly... data (shift midnight) does not make sense in our opinion.
But when user need more flexibility, then myClim Raw-format
can be used, In Raw-format \code{use_utc} is not limited, user can shift an data without the restrictions.
See \link{myClim-package}}

\item{show_logger_name}{if TRUE, the logger name is included in the column name (default FALSE)}
}
\value{
data.frame with columns:
\itemize{
\item datetime
\item locality1_sensor1
\item ...
\item ...
\item localityN_sensorN
}
}
\description{
This function converts myClim object to the R data.frame with values of sensor in wide format.
}
\details{
First column of the output data.frame is datetime followed by the
columns for every sensor. When reshaping uncleaned data in Raw-format, only one logger
per myClim object is allowed to avoid potential confusion in case, there are
duplicated values, allowed in uncleaned Raw myClim format.
Name of the column is in format:
\itemize{
\item \code{localityid_loggerid_serialnumber_sensorname} for Raw-format and \code{show_logger_name=FALSE}
\item \code{localityid_loggername_sensorname}for Raw-format and \code{show_logger_name=TRUE}
\item \code{localityid_sensorname} for Agg-format
}

The less complex wide table is returned when exporting single sensor across localities.
}
\examples{
example_tms_wideformat <- mc_reshape_wide(mc_data_example_clean, c("A6W79", "A2E32"),
                                          c("TMS_T1", "TMS_T2"))
}
