% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{mc_reshape_wideformat}
\alias{mc_reshape_wideformat}
\title{Wideformat of sensor values}
\usage{
mc_reshape_wideformat(data, localities = c(), sensors = c())
}
\arguments{
\item{data}{all data in standard format}

\item{localities}{names of localities; if empty then all}

\item{sensors}{names of sensors; if empty then all}
}
\value{
data in standard format
}
\description{
This function create data.frame with values of sensor in wide format.
}
\examples{
example_tms_t1_table <- microclim::mc_reshape_wideformat(example_tms_data, c("LOC_1", "LOC_2"), c("T1", "T2"))
}
