% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{mc_save}
\alias{mc_save}
\title{Save myClim object}
\usage{
mc_save(data, file)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{file}{path to output .rds file}
}
\description{
This function was designed for saving the myClim data object to an
.rds file, which can be later correctly loaded by any further version
of myClim package with \link{mc_load}. This is the safest way how to store and
share your myClim data.
}
\examples{
\dontrun{mc_save("myClim_data.rds")}
}
