% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{mc_save}
\alias{mc_save}
\title{Save myClim object}
\usage{
mc_save(data, file)
}
\arguments{
\item{data}{myClim object see \link{myClim-package}}

\item{file}{path to output .rds file}
}
\value{
RDS file saved at the output path destination
}
\description{
This function was designed for saving the myClim data object to an
.rds file, which can be later correctly loaded by any further version
of myClim package with \link{mc_load}. This is the safest way how to store and
share your myClim data.
}
\examples{
tmp_dir <- tempdir()
tmp_file <- tempfile(tmpdir = tmp_dir)
mc_save(mc_data_example_agg, tmp_file)
file.remove(tmp_file)
}
