% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_Sensor-class}
\alias{mc_Sensor-class}
\alias{mc_Sensor}
\title{Class for sensor definition}
\description{
Sensor definitions are stored in \link{mc_data_sensors}.
}
\section{Slots}{

\describe{
\item{\code{sensor_id}}{unique identifier of sensor (TMS_T1, TMS_T2, TMS_T3, TMS_TMSmoisture, ...)}

\item{\code{logger}}{name of logger (TMS, ThermoDatalogger, ...)}

\item{\code{physical}}{unit of sensor (T_C, TMSmoisture, moisture, RH_perc) (default NA)}

\item{\code{description}}{character info}

\item{\code{value_type}}{type of values (real, integer, logical) (default real)}

\item{\code{min_value}}{minimal value (default NA)}

\item{\code{max_value}}{maximal value (default NA)}

\item{\code{plot_color}}{color in pot (default "")}

\item{\code{plot_line_width}}{width of line in plot (default 1)}
}}

\seealso{
\link{mc_data_sensors}
}
