% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_TOMSTDataFormat-class}
\alias{mc_TOMSTDataFormat-class}
\alias{mc_TOMSTDataFormat}
\title{Class for reading Tomst logger files}
\description{
Provides the key for the column in source files. Where is the date,
in what format is the date, in which columns are records of which sensors.
The code defining the class is in section methods ./R/model.R
}
\seealso{
\link{mc_DataFormat}, \link{mc_data_formats}, \linkS4class{mc_TOMSTJoinDataFormat}
}
