% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{mc_TOMSTJoinDataFormat-class}
\alias{mc_TOMSTJoinDataFormat-class}
\alias{mc_TOMSTJoinDataFormat}
\title{Class for reading TMS join files}
\description{
Provides the key for the column in source files. Where is the date,
in what format is the date, in which columns are records of which sensors.
The code defining the class is in section methods ./R/model.R
}
\details{
TMS join file format is the output of IBOT internal post-processing of Tomst logger files.
}
\seealso{
\link{mc_DataFormat},\link{mc_data_formats},\linkS4class{mc_TOMSTDataFormat}, \linkS4class{mc_TOMSTJoinDataFormat}
}
