% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclim-package.R
\docType{package}
\name{myClim-package}
\alias{myClim}
\alias{myClim-package}
\title{myClim: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
