% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myClim-package.R
\docType{package}
\name{myClim-package}
\alias{myClim}
\alias{myClim-package}
\title{myClim: R package for processing microclimatic data}
\description{
Package myClim was designed for handling the microclimatic data in R.
myClim workflow begin at the reading data primary from microclimatic dataloggers,
but can be also reading of meteorological station data from files.
Cleaning time step, time zone settings and metadata collecting is the next step of the work flow.
With myClim tools one can crop, join, downscale, and convert microclimatic
data formats, sort them into localities, request descriptive characteristics
and compute microclimatic variables. Handy plotting functions are
provided with smart defaults.

\strong{myClim object}

We implemented two slightly different data formats of myClim objects calling them:
Prep-format and Calc-format.
Prep-format is designed for data preparation. Mainly data cleaning,
metadata gathering, time zones handling and multiple downloading missions
from single localities joining. Output myClim object from
read functions \code{\link[=mc_read_files]{mc_read_files()}} and \code{\link[=mc_read_data]{mc_read_data()}} is a Prep-format.
Prep-format contains information about localities, loggers and sensors.
Function \code{\link[=mc_agg]{mc_agg()}} converts data from Prep-format to Calc-format.
Calc-format is designed mainly for calculations,analysis and virtual
loggers derivations on the basis of cleaned, final microclimatic data.
Calc-formt has not the level of loggers and their metadata.
In Calc-format sensors are organized directly in localities, without loggers.

Microclimatic data are stored in myClim objects in regular R structure
consists of classes and lists slightly resembling database scheme.
The top level of the structure is the \code{locality}. It has on metadata e.g.
altitude, latitude or user data  dedicated for any type of metadata user need.
For detail desription of locality metadata see \link{mc_LocalityMetadata}.
On the locality in Prep-format there are \code{loggers} or in Calc-format directly \code{sensors}.
See below. Loggers represents the files imported with myClim reading functions.
Both \code{loggesr} and \code{sensors} have own metadata. One logger could host more sensors e.g.
TOMST TMS logger hosing TMS_T1 soil temperature, TMS_T2 surface temperature,
TMS_T3 air temperature , TMS_TMSmoisture soil moisture; or one logger could host
single sensor.
For detailed description of logger and sensor metadata
see \link{mc_LoggerMetadata}, \link{mc_SensorMetadata}, \link{mc_data_sensors}

\strong{Prep-format}

Is the output of reading functions \code{\link[=mc_read_files]{mc_read_files()}} and \code{\link[=mc_read_data]{mc_read_data()}}.
Prep-format is mainly designed for preparation. Especially for the situation when
user downloads the logger during more field visits in time. E. g. We have the logger
on the locality measuring 10 years. We download data twice a year and at the end we
wan to work with single correctly joined time series of the 10 years records.
For this type of time joining it is useful to have a level "logger" in myClim
object which allows to keep separate logger  downloads and prepare them for joining.
\emph{Joining is not implemented yet in myClim. We are working on it.}

Within the logger all sensors share time series. In Calc-format where level of
logger is missing, all sensors within the locality sahre time series. In Prep-format
the time series between the logger or between the localities can be of different time
step. E.g. on the locality there can be one logger measuring in time step
15 minutes and another one measuring once a day.  In Calc-format this is not allowed.
Therefore it is necessary to use \code{\link[=mc_agg]{mc_agg()}} to switch to the Calc-format.
Later there will be join function implemented also switching from Prep-format to Calc-format.

Schema of myClim \strong{Prep-format}

\preformatted{
+-------------------------------------------------------------------------------+
| locality[1]                                                                   |
+-------------------------------------------------------------------------------+
|           +-----------------------------+                                     |
| $metadata | mc_LocalityMetadata - class |                                     |
|           +-----------------------------+                                     |
|           | @locality_id                |                                     |
|           | @altitude                   |                                     |
|           | @lat_wgs84                  |                                     |
|           | @lon_wgs84                  |                                     |
|           | @tz_offset                  |                                     |
|           | @tz_type                    |                                     |
|           | @user_data                  |                                     |
|           +-----------------------------+                                     |
|           +------------------------------------------------------------------+|
| $loggers  | logger[1]                                                        ||
|           +------------------------------------------------------------------+|
|           |             +---------------------------+                        ||
|           | $metadata   | mc_LoggerMetadata - class |                        ||
|           |             +---------------------------+                        ||
|           |             | @type                     |                        ||
|           |             | @serial_number            |                        ||
|           |             +---------------------------+                        ||
|           |             +----------------------------+                       ||
|           | $clean_info | mc_LoggerCleanInfo - class |                       ||
|           |             +----------------------------+                       ||
|           |             | @step                      |                       ||
|           |             | @count_duplicits           |                       ||
|           |             | @count_missed              |                       ||
|           |             | @count_disordered          |                       ||
|           |             | @rounded                   |                       ||
|           |             +----------------------------+                       ||
|           | $datetime   POSIXct vector                                       ||
|           |             +---------------------------------------------------+||
|           | $sensors    | sensor[1]                                         |||
|           |             +---------------------------------------------------+||
|           |             |              +---------------------------+        |||
|           |             | $metadata    | mc_SensorMetadata - class |        |||
|           |             |              +---------------------------+        |||
|           |             |              | @sensor_id                |        |||
|           |             |              | @name                     |        |||
|           |             |              | @height                   |        |||
|           |             |              | @calibrated               |        |||
|           |             |              +---------------------------+        |||
|           |             |              +----------+------------+-----------+|||
|           |             | $calibration | datetime | cor_factor | cor_slope ||||
|           |             |              | ...      | ...        | ...       ||||
|           |             |              +----------+------------+-----------+|||
|           |             | $values      numeric/logical vector               |||
|           |             |              +-----+-------+-----+-------+        |||
|           |             | $states      | tag | start | end | value |        |||
|           |             |              | ... | ...   | ... | ...   |        |||
|           |             |              +-----+-------+-----+-------+        |||
|           |             +---------------------------------------------------+||
|           |             +---------------------------------------------------+||
|           |             | sensor[2]                                         |||
|           |             +---------------------------------------------------+||
|           |             ...                                                  ||
|           |             +---------------------------------------------------+||
|           |             | sensor[n]                                         |||
|           |             +---------------------------------------------------+||
|           +------------------------------------------------------------------+|
|           +------------------------------------------------------------------+|
|           | logger[2]                                                        ||
|           +------------------------------------------------------------------+|
|           ...                                                                 |
|           +------------------------------------------------------------------+|
|           | logger[n]                                                        ||
|           +------------------------------------------------------------------+|
+-------------------------------------------------------------------------------+
+-------------------------------------------------------------------------------+
| locality[2]                                                                   |
+-------------------------------------------------------------------------------+
...
+-------------------------------------------------------------------------------+
| locality[n]                                                                   |
+-------------------------------------------------------------------------------+}

\strong{Calc-format}

Is the output of \code{\link[=mc_agg]{mc_agg()}} function.
Calc-format is mainly designed for calculations with clean, nicely
organized microclimatic data with unified time step, supplemented with necessary metadata.
User can calculate various "virtual sensors" on the locality using existing ones
e. g. calculate volumetric water content form TMS raw moisture records see \link{mc_calc_vwc};
calculate growing degree days or freezing degree days see \link{mc_calc_gdd}, \link{mc_calc_fdd};
or estimate snow presence from near ground temperature \link{mc_calc_snow}.
\emph{Only few calc functions are available do far, we are working on it.}

In Calc-format the level of logger is missing. This is in contrast with Prep-format.
In Calc-format sensors are organized directly in the localities.
Within the locality of Calc-format all sensors share time series (start, end, step)
all localities shares the same time step, can have different start, end.
This is  practical for calculations, and it is achieved using \code{\link[=mc_agg]{mc_agg()}} function.
E.g. we have 10 localities, each of them hosts several sensors.
The same as in Prep-format \code{locality} and \code{sensor} has own metadata
see \link{mc_SensorMetadata}, \link{mc_data_sensors}, \link{mc_LocalityMetadata}.

In contrast with Prep-format the Calc-formated myClim objects have additional
metadata on the level of  whole object \linkS4class{mc_MainMetadata} holding information
on time step which has whole object. Step is stored in minutes (\code{data$metadata@step}).
But some steps can not be represented by minutes. For example step \code{month} has variable
number of minutes. Therefore, metadata contains also text
representation of the step (\code{data$metadata@step_text}). Common time step
of myClim Calc-format objects is practical when merging multiple object. See \link{mc_prep_merge}

Schema of myClim \strong{Calc-format}

\preformatted{
            +-------------------------+
$metadata   | mc_MainMetadata - class |
            +-------------------------+
            | @step                   |
            | @step_text              |
            +-------------------------+
            +-----------------------------------------------------------------+
$localities | locality[1]                                                     |
            +-----------------------------------------------------------------+
            |            +-----------------------------+                      |
            | $metadata  | mc_LocalityMetadata - class |                      |
            |            +-----------------------------+                      |
            |            | @locality_id                |                      |
            |            | @altitude                   |                      |
            |            | @lat_wgs84                  |                      |
            |            | @lon_wgs84                  |                      |
            |            | @tz_offset                  |                      |
            |            | @tz_type                    |                      |
            |            | @user_data                  |                      |
            |            +-----------------------------+                      |
            | $datetime  POSIXct vector                                       |
            |            +---------------------------------------------------+|
            | $sensors   | sensor[1]                                         ||
            |            +---------------------------------------------------+|
            |            |              +---------------------------+        ||
            |            | $metadata    | mc_SensorMetadata - class |        ||
            |            |              +---------------------------+        ||
            |            |              | @sensor_id                |        ||
            |            |              | @name                     |        ||
            |            |              | @height                   |        ||
            |            |              | @calibrated               |        ||
            |            |              +---------------------------+        ||
            |            |              +----------+------------+-----------+||
            |            | $calibration | datetime | cor_factor | cor_slope |||
            |            |              | ...      | ...        | ...       |||
            |            |              +----------+------------+-----------+||
            |            | $values      numeric/logical vector               ||
            |            |              +-----+-------+-----+-------+        ||
            |            | $states      | tag | start | end | value |        ||
            |            |              | ... | ...   | ... | ...   |        ||
            |            |              +-----+-------+-----+-------+        ||
            |            +---------------------------------------------------+|
            |            +---------------------------------------------------+|
            |            | sensor[2]                                         ||
            |            +---------------------------------------------------+|
            |            ...                                                  |
            |            +---------------------------------------------------+|
            |            | sensor[n]                                         ||
            |            +---------------------------------------------------+|
            +-----------------------------------------------------------------+
            +-----------------------------------------------------------------+
            | locality[2]                                                     |
            +-----------------------------------------------------------------+
            ...
            +-----------------------------------------------------------------+
            | locality[n]                                                     |
            +-----------------------------------------------------------------+}
}
\author{
\strong{Maintainer}: Vojtěch Kalčík \email{Vojtech.Kalcik@ibot.cas.cz}

Authors:
\itemize{
  \item Matěj Man \email{Matej.Man@ibot.cas.cz}
  \item Martin Macek \email{Martin.Macek@ibot.cas.cz}
  \item Jan Wild \email{Jan.Wild@ibot.cas.cz}
  \item Martin Kopecký \email{Martin.Kopecky@ibot.cas.cz}
  \item Josef Brůna \email{Josef.Bruna@ibot.cas.cz}
  \item Lucia Hederová \email{Lucia.Hederova@ibot.cas.cz}
}

Other contributors:
\itemize{
  \item Institute of Botany of the Czech Academy of Sciences [copyright holder]
}

}
\keyword{internal}
