% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myClim-package.R
\docType{package}
\name{myClim-package}
\alias{myClim}
\alias{myClim-package}
\title{myClim: R package for processing microclimatic data}
\description{
Package myClim use data in two formats. Names of formats are prep-format and calc-format.
Loaded data from source files by read functions are in prep-format. It Contains information about localities,
loggers and sensors. Function mc_agg converts data to calc-format. Calc-format doesn't contain loggers metadata.
Sensors are organized in localities. Calc-format is focused to analyzing data and calculation new virtual sensors.

Prep-format
\preformatted{
+--------------------------------------------------------------------------+
| locality[1]                                                              |
+--------------------------------------------------------------------------+
|           +-----------------------------+                                |
| $metadata | mc_LocalityMetadata - class |                                |
|           +-----------------------------+                                |
|           | @locality_id                |                                |
|           | @altitude                   |                                |
|           | @lat_wgs84                  |                                |
|           | @lon_wgs84                  |                                |
|           | @tz_offset                  |                                |
|           | @tz_type                    |                                |
|           | @user_data                  |                                |
|           +-----------------------------+                                |
|           +-------------------------------------------------------------+|
| $loggers  | logger[1]                                                   ||
|           +-------------------------------------------------------------+|
|           |             +---------------------------+                   ||
|           | $metadata   | mc_LoggerMetadata - class |                   ||
|           |             +---------------------------+                   ||
|           |             | @type                     |                   ||
|           |             | @serial_number            |                   ||
|           |             +---------------------------+                   ||
|           |             +----------------------------+                  ||
|           | $clean_info | mc_LoggerCleanInfo - class |                  ||
|           |             +----------------------------+                  ||
|           |             | @step                      |                  ||
|           |             | @count_duplicits           |                  ||
|           |             | @count_missed              |                  ||
|           |             | @count_disordered          |                  ||
|           |             +----------------------------+                  ||
|           | $datetime   POSIXct vector                                  ||
|           |             +----------------------------------------------+||
|           | $sensors    | sensor[1]                                    |||
|           |             +----------------------------------------------+||
|           |             |              +---------------------------+   |||
|           |             | $metadata    | mc_SensorMetadata - class |   |||
|           |             |              +---------------------------+   |||
|           |             |              | @sensor_id                |   |||
|           |             |              | @name                     |   |||
|           |             |              | @height                   |   |||
|           |             |              | @calibrated               |   |||
|           |             |              +---------------------------+   |||
|           |             |              +----------+-------+-----------+|||
|           |             | $calibration | datetime | slope | intercept ||||
|           |             |              | ...      | ...   | ...       ||||
|           |             |              +----------+-------+-----------+|||
|           |             | $values      numeric/logical vector          |||
|           |             |              +-----+-------+-----+           |||
|           |             | $states      | tag | start | end |           |||
|           |             |              | ... | ...   | ... |           |||
|           |             |              +-----+-------+-----+           |||
|           |             +----------------------------------------------+||
|           |             +----------------------------------------------+||
|           |             | sensor[2]                                    |||
|           |             +----------------------------------------------+||
|           |             ...                                             ||
|           |             +----------------------------------------------+||
|           |             | sensor[n]                                    |||
|           |             +----------------------------------------------+||
|           +-------------------------------------------------------------+|
|           +-------------------------------------------------------------+|
|           | logger[2]                                                   ||
|           +-------------------------------------------------------------+|
|           ...                                                            |
|           +-------------------------------------------------------------+|
|           | logger[n]                                                   ||
|           +-------------------------------------------------------------+|
+--------------------------------------------------------------------------+
+--------------------------------------------------------------------------+
| locality[2]                                                              |
+--------------------------------------------------------------------------+
...
+--------------------------------------------------------------------------+
| locality[n]                                                              |
+--------------------------------------------------------------------------+}
}
\keyword{internal}
