% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare.read_files}
\alias{prepare.read_files}
\title{Reading files}
\usage{
prepare.read_files(files, logger_type)
}
\arguments{
\item{files}{vector of character - files with data}

\item{logger_type}{character - type of logger (TMS)}
}
\value{
data in standard format
}
\description{
This function read data files of one logger type. Locality is set None.
}
