% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare.read_logger}
\alias{prepare.read_logger}
\title{Logger raw data reading}
\usage{
prepare.read_logger(
  filename,
  serial_number,
  id_locality,
  data_format,
  tz = "UTC"
)
}
\arguments{
\item{filename}{path to source file}

\item{serial_number}{serial number of logger}

\item{id_locality}{id of locality}

\item{data_format}{definition of format source file}

\item{tz}{time zone of dates in source data}
}
\value{
data in standard format
}
\description{
This function read raw data of logger and return data in standard format.
}
