% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{print.myClimList}
\alias{print.myClimList}
\title{Print function for myClim object}
\usage{
\method{print}{myClimList}(x, ...)
}
\arguments{
\item{x}{myClim object see \link{myClim-package}}

\item{...}{other parameters from function print for tibble \link[tibble:tibble]{tibble::tibble}}
}
\description{
Function print metadata of myClim object and table from function mc_info().
}
\examples{
print(mc_data_example_agg, n=10)
}
